% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{addEntry}
\alias{addEntry}
\title{Add entry to list.}
\usage{
addEntry(object, entry)
}
\arguments{
\item{object}{generic object}

\item{entry}{entry to add}
}
\value{
generic object
}
\description{
Add entry to list.
}
