% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/scenario_action.R
\name{apply_action}
\alias{apply_action}
\alias{apply_action,campsis_model,replace_action-method}
\alias{apply_action,dataset,replace_action-method}
\title{Apply some action on the given object.}
\usage{
apply_action(object, action)

\S4method{apply_action}{campsis_model,replace_action}(object, action)

\S4method{apply_action}{dataset,replace_action}(object, action)
}
\arguments{
\item{object}{any object}

\item{action}{action to apply}
}
\value{
updated object
}
\description{
Apply some action on the given object.
}
