% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{apply_compartment_characteristics}
\alias{apply_compartment_characteristics}
\title{Apply compartment characteristics from model.
In practice, only compartment infusion duration needs to be applied.}
\usage{
apply_compartment_characteristics(table, properties)
}
\arguments{
\item{table}{current dataset}

\item{properties}{compartment properties from model}
}
\value{
updated dataset
}
\description{
Apply compartment characteristics from model.
In practice, only compartment infusion duration needs to be applied.
}
