% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/outfun.R, R/outfuns.R
\name{apply_outfun}
\alias{apply_outfun}
\alias{apply_outfun,outfun-method}
\alias{apply_outfun,outfuns-method}
\title{Apply output function(s) on the given Campsis results.}
\usage{
apply_outfun(x, outfun, level = NULL, ...)

\S4method{apply_outfun}{outfun}(x, outfun, level = NULL, ...)

\S4method{apply_outfun}{outfuns}(x, outfun, level = NULL, ...)
}
\arguments{
\item{x}{Campsis simulation results}

\item{outfun}{output function(s), an `outfun` or `outfuns` object}

\item{level}{level at which the output function is applied, 'replicate' by default}

\item{...}{extra arguments transmitted automatically by Campsis (e.g. `replicate` for the replicate number)}
}
\value{
the updated Campsis results
}
\description{
Apply output function(s) on the given Campsis results.
}
