% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{applyAction}
\alias{applyAction}
\title{Apply some action on the given object.}
\usage{
applyAction(object, action)
}
\arguments{
\item{object}{any object}
}
\value{
updated object
}
\description{
Apply some action on the given object.
}
