% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{applyCompartmentCharacteristics}
\alias{applyCompartmentCharacteristics}
\title{Apply compartment characteristics from model.
In practice, only compartment infusion duration needs to be applied.}
\usage{
applyCompartmentCharacteristics(table, properties)
}
\arguments{
\item{table}{current dataset}

\item{properties}{compartment properties from model}
}
\value{
updated dataset
}
\description{
Apply compartment characteristics from model.
In practice, only compartment infusion duration needs to be applied.
}
