% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{applyScenario}
\alias{applyScenario}
\title{Apply scenario to the given model or dataset.}
\usage{
applyScenario(x, scenario)
}
\arguments{
\item{x}{the given model or dataset}

\item{scenario}{the scenario to be applied}
}
\value{
an updated model or dataset
}
\description{
Apply scenario to the given model or dataset.
}
\keyword{internal}
