% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.R
\docType{class}
\name{arm-class}
\alias{arm-class}
\title{Arm class.}
\description{
Arm class.
}
\section{Slots}{

\describe{
\item{\code{id}}{arm unique ID, integer}

\item{\code{subjects}}{number of subjects in arm, integer}

\item{\code{protocol}}{protocol}

\item{\code{covariates}}{covariates}
}}

