% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_vector.R
\name{as.numeric,time_sequence-method}
\alias{as.numeric,time_sequence-method}
\title{Time sequence to numeric vector.}
\usage{
\S4method{as.numeric}{time_sequence}(x)
}
\arguments{
\item{x}{time sequence object}
}
\value{
a numeric vector
}
\description{
Time sequence to numeric vector.
}
