% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_vector.R
\name{as.numeric,time_vector-method}
\alias{as.numeric,time_vector-method}
\title{Time vector to numeric vector.}
\usage{
\S4method{as.numeric}{time_vector}(x)
}
\arguments{
\item{x}{time vector object}
}
\value{
a numeric vector
}
\description{
Time vector to numeric vector.
}
