% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{assign_dose_number}
\alias{assign_dose_number}
\title{Assign dose number to each treatment entry.}
\usage{
assign_dose_number(object)
}
\arguments{
\item{object}{treatment}
}
\value{
updated treatment object
}
\description{
Assign dose number to each treatment entry.
}
\keyword{internal}
