% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{assignDoseNumber}
\alias{assignDoseNumber}
\title{Assign dose number to each treatment entry.}
\usage{
assignDoseNumber(object)
}
\arguments{
\item{object}{treatment}
}
\value{
updated treatment object
}
\description{
Assign dose number to each treatment entry.
}
