% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{BinomialDistribution}
\alias{BinomialDistribution}
\title{Binomial distribution.}
\usage{
BinomialDistribution(trials, prob)
}
\arguments{
\item{trials}{number of Bernoulli trials per observation (=subject), integer}

\item{prob}{probability of success for each trial}
}
\value{
a binomial distribution
}
\description{
Binomial distribution.
}
