% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioavailability.R
\name{Bioavailability}
\alias{Bioavailability}
\title{Create a bioavailability for the specified compartment.}
\usage{
Bioavailability(compartment, distribution)
}
\arguments{
\item{compartment}{compartment number}

\item{distribution}{distribution}
}
\value{
bioavailability
}
\description{
Create a bioavailability for the specified compartment.
}
