% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_entry.R
\name{Bolus}
\alias{Bolus}
\title{Create one or several bolus(es).}
\usage{
Bolus(
  time,
  amount,
  compartment = NULL,
  f = NULL,
  lag = NULL,
  ii = NULL,
  addl = NULL,
  wrap = FALSE,
  ref = NULL
)
}
\arguments{
\item{time}{treatment time(s), numeric value or vector. First treatment time if used together with ii and addl.}

\item{amount}{amount to give as bolus, single numeric value}

\item{compartment}{compartment index or name to give the bolus(es). A vector of integers or names can be used for a complex model administration.}

\item{f}{fraction of dose amount, distribution}

\item{lag}{dose lag time, distribution}

\item{ii}{interdose interval, requires argument 'time' to be a single numeric value}

\item{addl}{number of additional doses, requires argument 'time' to be a single integer value}

\item{wrap}{if TRUE, the bolus wrapper will be stored as is in the dataset, otherwise,
it will be split into a list of infusions distinct in time. Default is FALSE.}

\item{ref}{wrapper reference, used to identify the wrapper in the dataset, single character value}
}
\value{
a single bolus or a list of boluses
}
\description{
Create one or several bolus(es).
}
