% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\docType{class}
\name{bootstrap-class}
\alias{bootstrap-class}
\title{Bootstrap class.}
\description{
Bootstrap class.
}
\section{Slots}{

\describe{
\item{\code{data}}{data frame to be bootstrapped. Column 'BS_ID' is mandatory and
corresponds to the original row ID from the bootstrap. It must be numeric and unique.
Other columns are covariates to be bootstrapped (row by row).}

\item{\code{replacement}}{values can be reused or not, logical}

\item{\code{random}}{values are drawn randomly, logical}

\item{\code{export_id}}{tell Campsis if 'BS_ID' must be exported into the dataset, logical}
}}

