% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\docType{class}
\name{bootstrap_distribution-class}
\alias{bootstrap_distribution-class}
\title{Bootstrap distribution class.}
\description{
Bootstrap distribution class.
}
\section{Slots}{

\describe{
\item{\code{data}}{values to draw, numeric vector}

\item{\code{replacement}}{values can be reused or not, logical}

\item{\code{random}}{values are drawn randomly, logical}
}}

