% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{BootstrapCovariate}
\alias{BootstrapCovariate}
\title{Create a bootstrap covariate. During covariate sampling, PMXsim will generate
values depending on the given data and arguments.}
\usage{
BootstrapCovariate(name, data, replacement = FALSE, random = FALSE)
}
\arguments{
\item{name}{covariate name, character}

\item{data}{values to draw, numeric vector}

\item{replacement}{values can be reused or not, logical}

\item{random}{values are drawn randomly, logical}
}
\value{
a covariate
}
\description{
Create a bootstrap covariate. During covariate sampling, PMXsim will generate
values depending on the given data and arguments.
}
