% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campsis_metadata.R
\docType{class}
\name{campsis_metadata-class}
\alias{campsis_metadata-class}
\title{Campsis metadata class.}
\description{
Campsis metadata class.
}
\section{Slots}{

\describe{
\item{\code{dataset}}{Campsis dataset object}

\item{\code{dest}}{destination simulation engine}

\item{\code{scenarios}}{Campsis scenarios object}

\item{\code{outvars}}{character vector of output variable names}

\item{\code{outfun}}{collection of output functions}

\item{\code{replicates}}{number of replicates}
}}

