% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{campsis_tbl-class}
\alias{campsis_tbl-class}
\alias{campsis_tbl}
\title{Campsis table class (see this class as an interface)}
\description{
Campsis table class (see this class as an interface)
}
