% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_entry.R
\name{check_ii_and_addl}
\alias{check_ii_and_addl}
\title{Check ii and addl arguments in addition to time.}
\usage{
check_ii_and_addl(time, ii, addl)
}
\arguments{
\item{time}{treatment time(s)}

\item{ii}{interdose interval}

\item{addl}{number of additional doses}
}
\value{
no return value
}
\description{
Check ii and addl arguments in addition to time.
}
\keyword{internal}
