% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_entry.R
\name{checkIIandADDL}
\alias{checkIIandADDL}
\title{Check ii and addl arguments in addition to time.}
\usage{
checkIIandADDL(time, ii, addl)
}
\arguments{
\item{time}{treatment time(s)}

\item{ii}{interdose interval}

\item{addl}{number of additional doses}
}
\description{
Check ii and addl arguments in addition to time.
}
\keyword{internal}
