% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun_core.R
\name{compute_stats}
\alias{compute_stats}
\title{Compute generic statistics over time.}
\usage{
compute_stats(
  x,
  variable,
  strata = get_default_strata(),
  stats = c("p5", "median", "p95")
)
}
\arguments{
\item{x}{data frame}

\item{variable}{variable(s) used to compute the statistics, character vector.}

\item{strata}{named vector with the strata to use, default is c(SCENARIO="all", ARM="all").}

\item{stats}{character vector of statistics to compute. Supported: "median", "mean", or percentiles like "p5", "p95", "p2.5", "p97.5".}
}
\value{
a summary table in long format
}
\description{
Compute generic statistics over time.
}
