% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_progress.R
\name{computeIncrementalProgress}
\alias{computeIncrementalProgress}
\title{Compute incremental progress.}
\usage{
computeIncrementalProgress(object, no_slices = TRUE)
}
\arguments{
\item{object}{simulation progress object}

\item{no_slices}{don't take slices into account in the incremental progress}
}
\value{
incremental progress, in percentage
}
\description{
Compute incremental progress.
}
\keyword{internal}
