% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\docType{class}
\name{constant_distribution-class}
\alias{constant_distribution-class}
\title{Constant distribution class.}
\description{
Constant distribution class.
}
\section{Slots}{

\describe{
\item{\code{value}}{covariate value, single numeric value}
}}

