% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{ConstantCovariate}
\alias{ConstantCovariate}
\title{Create a constant covariate. Its value will be constant across all arms and subjects.}
\usage{
ConstantCovariate(name, value)
}
\arguments{
\item{name}{covariate name, character}

\item{value}{covariate value, numeric}
}
\value{
a covariate
}
\description{
Create a constant covariate. Its value will be constant across all arms and subjects.
}
