% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{ConstantDistribution}
\alias{ConstantDistribution}
\title{Create a constant distribution. Its value will be constant across all generated samples.}
\usage{
ConstantDistribution(value)
}
\arguments{
\item{value}{covariate value, single numeric value}
}
\value{
a constant distribution (same value for all samples)
}
\description{
Create a constant distribution. Its value will be constant across all generated samples.
}
