% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{contains}
\alias{contains}
\title{Check if an element exists in list.}
\usage{
contains(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{element to check if exists}
}
\value{
logical value
}
\description{
Check if an element exists in list.
}
