% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{convert}
\alias{convert}
\title{Convert generic object according to given configuration.}
\usage{
convert(object, config)
}
\arguments{
\item{object}{generic object}

\item{config}{specified configuration}
}
\value{
conversion output
}
\description{
Convert generic object according to given configuration.
}
