% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utilities.R
\name{convert_time}
\alias{convert_time}
\title{Convert numeric time vector based on the provided units.}
\usage{
convert_time(x, from, to)
}
\arguments{
\item{x}{numeric time vector}

\item{from}{unit of x, single character value}

\item{to}{destination unit, single character value}
}
\value{
numeric vector with the converted times
}
\description{
Convert numeric time vector based on the provided units.
}
