% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{counterBalanceLocfMode}
\alias{counterBalanceLocfMode}
\title{Counter-balance LOCF mode for occasions & IOV.
This function will simply shift all the related occasion & IOV columns to the left (by one).}
\usage{
counterBalanceLocfMode(table, columnNames)
}
\arguments{
\item{table}{current table}

\item{columnNames}{columns to be counter-balanced}
}
\value{
2-dimensional dataset
}
\description{
Counter-balance LOCF mode for occasions & IOV.
This function will simply shift all the related occasion & IOV columns to the left (by one).
}
\keyword{internal}
