% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{counterBalanceNocbMode}
\alias{counterBalanceNocbMode}
\title{Counter-balance NOCB mode for occasions & IOV.
This function will simply shift all the related occasion & IOV columns to the right (by one).}
\usage{
counterBalanceNocbMode(table, columnNames)
}
\arguments{
\item{table}{current table}

\item{columnNames}{columns to be counter-balanced}
}
\value{
2-dimensional dataset
}
\description{
Counter-balance NOCB mode for occasions & IOV.
This function will simply shift all the related occasion & IOV columns to the right (by one).
}
\keyword{internal}
