% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\docType{class}
\name{covariate-class}
\alias{covariate-class}
\title{Covariate class.}
\description{
Covariate class.
}
\section{Slots}{

\describe{
\item{\code{name}}{covariate name, single character value}

\item{\code{distribution}}{covariate distribution}
}}

