% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{cutTableForEvent}
\alias{cutTableForEvent}
\title{Cut table for event starting at 'start' and ending at 'end'.}
\usage{
cutTableForEvent(table, start, end)
}
\arguments{
\item{table}{whole table, data frame}

\item{start}{event start time, numeric value}

\item{end}{event end time, numeric value}
}
\description{
Cut table for event starting at 'start' and ending at 'end'.
}
\keyword{internal}
