% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_schedule.R
\docType{class}
\name{cyclic_schedule-class}
\alias{cyclic_schedule-class}
\title{Cyclic schedule class.}
\description{
Cyclic schedule class.
}
\section{Slots}{

\describe{
\item{\code{duration}}{duration of the cycle, numeric value}

\item{\code{repetitions}}{number of additional repetitions to the base pattern, integer value}
}}

