% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_schedule.R
\name{CyclicSchedule}
\alias{CyclicSchedule}
\title{Cyclic schedule constructor.}
\usage{
CyclicSchedule(duration, repetitions)
}
\arguments{
\item{duration}{duration of the cycle, numeric value}

\item{repetitions}{number of additional repetitions to the base pattern, integer value}
}
\value{
a cyclic schedule
}
\description{
Cyclic schedule constructor.
}
