% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{class}
\name{dataset-class}
\alias{dataset-class}
\title{Dataset class.}
\description{
Dataset class.
}
\section{Slots}{

\describe{
\item{\code{arms}}{a list of treatment arms}

\item{\code{config}}{dataset configuration for export}

\item{\code{iiv}}{data frame containing the inter-individual variability (all ETAS) for the export}
}}

