% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_config.R
\docType{class}
\name{dataset_config-class}
\alias{dataset_config-class}
\title{Dataset config class.}
\arguments{
\item{defDepotCmt}{default depot compartment, integer}

\item{defObsCmt}{default observation compartment, integer}
}
\description{
Dataset config class.
}
