% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_summary.R
\name{DatasetSummary}
\alias{DatasetSummary}
\title{Create a dataset summary (internal method).}
\usage{
DatasetSummary()
}
\value{
a dataset summary
}
\description{
Create a dataset summary (internal method).
}
\keyword{internal}
