% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_settings.R
\name{Declare}
\alias{Declare}
\title{Create declare settings.}
\usage{
Declare(variables = character(0))
}
\arguments{
\item{variables}{uninitialized variables to be declared, only needed with mrgsolve}
}
\value{
Declare settings
}
\description{
Create declare settings.
}
