% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare_settings.R
\docType{class}
\name{declare_settings-class}
\alias{declare_settings-class}
\title{Declare settings class.}
\description{
Declare settings class.
}
\section{Slots}{

\describe{
\item{\code{variables}}{uninitialized variables to be declared, only needed with mrgsolve}
}}

