% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{default}
\alias{default}
\title{Get default element from list.}
\usage{
default(object, ...)
}
\arguments{
\item{object}{list object}

\item{...}{additional arguments}
}
\value{
the default element from list
}
\description{
Get default element from list.
}
