% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun.R
\name{DefaultOutfun}
\alias{DefaultOutfun}
\title{Default output function (identity function).}
\usage{
DefaultOutfun()
}
\value{
an output function that returns the Campsis results as is.
}
\description{
Default output function (identity function).
}
