% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{DiscreteDistribution}
\alias{DiscreteDistribution}
\title{Discrete distribution.}
\usage{
DiscreteDistribution(x, prob, replace = TRUE)
}
\arguments{
\item{x}{vector of one or more integers from which to choose}

\item{prob}{a vector of probability weights for obtaining the elements of the vector being sampled}

\item{replace}{should sampling be with replacement, default is TRUE}
}
\value{
a discrete distribution
}
\description{
Discrete distribution.
}
