% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_adaptation.R
\docType{class}
\name{dose_adaptation-class}
\alias{dose_adaptation-class}
\title{Dose adaptation class.}
\description{
Dose adaptation class.
}
\section{Slots}{

\describe{
\item{\code{formula}}{formula to apply, single character string, e.g. "AMT*WT"}

\item{\code{compartments}}{compartment numbers where the formula needs to be applied}
}}

