% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{dosingOnly}
\alias{dosingOnly}
\title{Filter CAMPSIS output on dosing rows.}
\usage{
dosingOnly(x)
}
\arguments{
\item{x}{data frame, CAMPSIS output}
}
\value{
a data frame with the dosing rows
}
\description{
Filter CAMPSIS output on dosing rows.
}
