% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_schedule.R
\name{DosingSchedule}
\alias{DosingSchedule}
\title{Dosing schedule constructor.}
\usage{
DosingSchedule()
}
\value{
a dosing schedule (schedule repeated at dose times)
}
\description{
Dosing schedule constructor.
}
