% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{drop_others}
\alias{drop_others}
\title{Return the 'DROP_OTHERS' string that may be used in the 'outvars' vector for
RxODE/mrgsolve to drop all others variables that are usually output in the resulting data frame.}
\usage{
drop_others()
}
\value{
a character value
}
\description{
Return the 'DROP_OTHERS' string that may be used in the 'outvars' vector for
RxODE/mrgsolve to drop all others variables that are usually output in the resulting data frame.
}
\keyword{internal}
