% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{EtaDistribution}
\alias{EtaDistribution}
\title{Create an ETA distribution. By default, the resulting distribution is a
normal distribution, computed as Normal(mean=0, sd=sqrt(OMEGA)).
It is currently not possible to change the distribution type.}
\usage{
EtaDistribution(omega)
}
\arguments{
\item{omega}{corresponding THETA name, character}
}
\value{
an ETA distribution
}
\description{
Create an ETA distribution. By default, the resulting distribution is a
normal distribution, computed as Normal(mean=0, sd=sqrt(OMEGA)).
It is currently not possible to change the distribution type.
}
