% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\docType{class}
\name{event-class}
\alias{event-class}
\title{Event class.}
\description{
Event class.
}
\section{Slots}{

\describe{
\item{\code{name}}{event name, character value}

\item{\code{times}}{interruption times, numeric vector}

\item{\code{fun}}{event function to apply at each interruption}

\item{\code{debug}}{output the variables that were changed through this event}
}}

