% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{Event}
\alias{Event}
\title{Create an interruption event.}
\usage{
Event(name = NULL, times, fun, debug = FALSE)
}
\arguments{
\item{name}{event name, character value}

\item{times}{interruption times, numeric vector}

\item{fun}{event function to apply at each interruption}

\item{debug}{output the variables that were changed through this event}
}
\value{
an event definition
}
\description{
Create an interruption event.
}
