% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{EventCovariate}
\alias{EventCovariate}
\title{Create an event covariate. These covariates can be modified further in
interruption events.}
\usage{
EventCovariate(name, distribution)
}
\arguments{
\item{name}{covariate name, character}

\item{distribution}{covariate distribution at time 0}
}
\value{
a time-varying covariate
}
\description{
Create an event covariate. These covariates can be modified further in
interruption events.
}
