% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_logic.R
\name{EventIteration}
\alias{EventIteration}
\title{Create an event iteration object.}
\usage{
EventIteration(index, start, end, inits = data.frame(), multiple = FALSE)
}
\arguments{
\item{index}{iteration index, starts at 1}

\item{start}{iteration start time}

\item{end}{iteration end time}

\item{inits}{initial values for all subjects, data frame}

\item{multiple}{TRUE if multiple iterations (i.e. simulation needs to stopped at least once), FALSE otherwise}
}
\value{
an event iteration object
}
\description{
Create an event iteration object.
}
\keyword{internal}
