% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{EventRelatedObservations}
\alias{EventRelatedObservations}
\title{Create an event-related observations list. Please note that the provided 'times' will
automatically be sorted. Duplicated times will be removed.}
\usage{
EventRelatedObservations(times, compartment = NA)
}
\arguments{
\item{times}{observation times, numeric vector}

\item{compartment}{compartment index, integer}
}
\value{
observations
}
\description{
Create an event-related observations list. Please note that the provided 'times' will
automatically be sorted. Duplicated times will be removed.
}
\keyword{internal}
