% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{export}
\alias{export}
\title{Export function.}
\usage{
export(object, dest, ...)
}
\arguments{
\item{object}{generic object}

\item{dest}{destination}

\item{...}{optional arguments}
}
\value{
specific object depending on given destination
}
\description{
Export function.
}
