% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{export_table_delegate}
\alias{export_table_delegate}
\title{Export table delegate.}
\usage{
export_table_delegate(model, dataset, dest, events, seed, tablefun, settings)
}
\arguments{
\item{model}{generic Campsis model}

\item{dataset}{Campsis dataset or 2-dimensional table}

\item{dest}{destination simulation engine, default is 'RxODE'}

\item{events}{interruption events}

\item{seed}{seed value}

\item{tablefun}{function or lambda formula to apply on exported 2-dimensional dataset}

\item{settings}{advanced simulation settings}
}
\value{
a data frame
}
\description{
Export table delegate.
}
\keyword{internal}
