% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{exportDelegate}
\alias{exportDelegate}
\title{Export delegate method. This method is common to RxODE and mrgsolve.}
\usage{
exportDelegate(object, dest, model, offset = 0)
}
\arguments{
\item{object}{current dataset}

\item{dest}{destination engine}

\item{model}{Campsis model, if provided, ETA's will be added to the dataset}
}
\value{
2-dimensional dataset, same for RxODE and mrgsolve
}
\description{
Export delegate method. This method is common to RxODE and mrgsolve.
}
\keyword{internal}
