% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{exportTableDelegate}
\alias{exportTableDelegate}
\title{Export table delegate.}
\usage{
exportTableDelegate(
  model,
  dataset,
  dest,
  events,
  seed,
  tablefun,
  nocb,
  nocbvars
)
}
\arguments{
\item{model}{generic CAMPSIS model}

\item{dataset}{CAMPSIS dataset or 2-dimensional table}

\item{dest}{destination simulation engine, default is 'RxODE'}

\item{events}{interruption events}

\item{seed}{seed value}

\item{tablefun}{function or lambda formula to apply on exported 2-dimensional dataset}

\item{nocb}{next-observation carried backward mode (NOCB), default value is TRUE for mrgsolve, FALSE for RxODE}
}
\value{
a data frame
}
\description{
Export table delegate.
}
\keyword{internal}
