% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{factorScenarios}
\alias{factorScenarios}
\title{Factor scenarios columns if not done yet.}
\usage{
factorScenarios(x, scenarios = NULL)
}
\arguments{
\item{x}{data frame}

\item{scenarios}{scenarios}
}
\description{
Factor scenarios columns if not done yet.
}
