% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{fillIOVOccColumns}
\alias{fillIOVOccColumns}
\title{Fill IOV/Occasion columns.}
\usage{
fillIOVOccColumns(table, columnNames, downDirectionFirst)
}
\arguments{
\item{table}{current table}

\item{columnNames}{the column names to fill}

\item{downDirectionFirst}{TRUE: first fill down then fill up (by ID & TIME). FALSE: First fill up (by ID & TIME), then fill down}
}
\value{
2-dimensional dataset
}
\description{
Problem in RxODE (LOCF mode) / mrgsolve (LOCF mode), if 2 rows have the same time (often: OBS then DOSE), first row covariate value is taken!
Workaround: identify these rows (group by ID and TIME) and apply a fill in the UP direction.
}
\keyword{internal}
