% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{filter}
\alias{filter}
\title{Filter.}
\usage{
filter(object, type)
}
\arguments{
\item{object}{generic object}

\item{type}{parameter type: theta, omega or sigma}
}
\value{
filtered object
}
\description{
Filter.
}
