% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\docType{class}
\name{fixed_distribution-class}
\alias{fixed_distribution-class}
\title{Fixed distribution class.}
\description{
Fixed distribution class.
}
\section{Slots}{

\describe{
\item{\code{values}}{covariate values, numeric vector (1 value per sample)}
}}

