% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{FixedCovariate}
\alias{FixedCovariate}
\title{Create a fixed covariate. Each subject will be assigned a fixed value, i.e.,
a value that does not vary over time.}
\usage{
FixedCovariate(name, values)
}
\arguments{
\item{name}{covariate name, character}

\item{values}{covariate values, numeric vector (1 value per subject)}
}
\value{
a covariate
}
\description{
Create a fixed covariate. Each subject will be assigned a fixed value, i.e.,
a value that does not vary over time.
}
