% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\docType{class}
\name{function_distribution-class}
\alias{function_distribution-class}
\title{Function distribution class.}
\description{
Function distribution class.
}
\section{Slots}{

\describe{
\item{\code{fun}}{function name, character (e.g. 'rnorm')}

\item{\code{args}}{list of arguments (e.g list(mean=70, sd=10))}
}}

