% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{FunctionCovariate}
\alias{FunctionCovariate}
\title{Create a function covariate. During covariate sampling, the provided function
will be responsible for generating values for each subject. If first argument
of this function is not the size (n), please tell which argument corresponds
to the size 'n' (e.g. list(size="n")).}
\usage{
FunctionCovariate(name, fun, args)
}
\arguments{
\item{name}{covariate name, character}

\item{fun}{function name, character (e.g. 'rnorm')}

\item{args}{list of arguments (e.g list(mean=70, sd=10))}
}
\value{
a covariate
}
\description{
Create a function covariate. During covariate sampling, the provided function
will be responsible for generating values for each subject. If first argument
of this function is not the size (n), please tell which argument corresponds
to the size 'n' (e.g. list(size="n")).
}
