% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{FunctionDistribution}
\alias{FunctionDistribution}
\title{Create a function distribution During covariate sampling, the provided function
will be responsible for generating values for each subject. If first argument
of this function is not the size (n), please tell which argument corresponds
to the size 'n' (e.g. list(size="n")).}
\usage{
FunctionDistribution(fun, args)
}
\arguments{
\item{fun}{function name, character (e.g. 'rnorm')}

\item{args}{list of arguments (e.g list(mean=70, sd=10))}

\item{name}{covariate name, character}
}
\value{
a covariate
}
\description{
Create a function distribution During covariate sampling, the provided function
will be responsible for generating values for each subject. If first argument
of this function is not the size (n), please tell which argument corresponds
to the size 'n' (e.g. list(size="n")).
}
