% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{generateIIV}
\alias{generateIIV}
\title{Generate IIV.}
\usage{
generateIIV(omega, n)
}
\arguments{
\item{omega}{omega matrix}

\item{n}{number of subjects}
}
\value{
IIV data frame
}
\description{
Generate IIV.
}
