% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utilities.R
\name{get_available_time_units}
\alias{get_available_time_units}
\title{Return the list of available time units.}
\usage{
get_available_time_units()
}
\value{
character vector
}
\description{
Return the list of available time units.
}
