% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_campsis_option}
\alias{get_campsis_option}
\title{Get Campsis option logic.}
\usage{
get_campsis_option(name, default)
}
\arguments{
\item{name}{option to search}

\item{default}{default value if option not found}
}
\value{
option value
}
\description{
Get Campsis option logic.
}
