% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_campsis_options}
\alias{get_campsis_options}
\title{Get the Campsis options (R options).}
\usage{
get_campsis_options()
}
\value{
global options for Campsis
}
\description{
Get the Campsis options (R options).
}
\keyword{internal}
