% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/covariates.R, R/arm.R, R/arms.R,
%   R/dataset.R
\name{get_fixed_covariates}
\alias{get_fixed_covariates}
\alias{get_fixed_covariates,covariates-method}
\alias{get_fixed_covariates,arm-method}
\alias{get_fixed_covariates,arms-method}
\alias{get_fixed_covariates,dataset-method}
\title{Get all fixed covariates.}
\usage{
get_fixed_covariates(object)

\S4method{get_fixed_covariates}{covariates}(object)

\S4method{get_fixed_covariates}{arm}(object)

\S4method{get_fixed_covariates}{arms}(object)

\S4method{get_fixed_covariates}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
all fixed covariates from object
}
\description{
Get all fixed covariates.
}
