% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_initial_conditions}
\alias{get_initial_conditions}
\title{Get initial conditions at simulation start-up.}
\usage{
get_initial_conditions(subdataset, iteration, cmtNames)
}
\arguments{
\item{subdataset}{subset of the dataset to simulate}

\item{iteration}{current iteration}

\item{cmtNames}{compartment names}
}
\value{
named numeric vector with the new initial conditions
}
\description{
Get initial conditions at simulation start-up.
}
\keyword{internal}
