% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/arm.R, R/arms.R, R/dataset.R
\name{get_occasions}
\alias{get_occasions}
\alias{get_occasions,arm-method}
\alias{get_occasions,arms-method}
\alias{get_occasions,dataset-method}
\title{Get all occasions.}
\usage{
get_occasions(object)

\S4method{get_occasions}{arm}(object)

\S4method{get_occasions}{arms}(object)

\S4method{get_occasions}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
all occasions from object
}
\description{
Get all occasions.
}
