% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{get_seed_for_parameters_sampling}
\alias{get_seed_for_parameters_sampling}
\title{Get seed for parameter uncertainty sampling.}
\usage{
get_seed_for_parameters_sampling(seed)
}
\arguments{
\item{seed}{original seed}
}
\value{
the seed value used to sample parameter uncertainty
}
\description{
Get seed for parameter uncertainty sampling.
}
\keyword{internal}
