% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_simulation_engine_type}
\alias{get_simulation_engine_type}
\title{Get simulation engine type.}
\usage{
get_simulation_engine_type(dest)
}
\arguments{
\item{dest}{destination engine, string form}
}
\value{
simulation engine type
}
\description{
Get simulation engine type.
}
\keyword{internal}
