% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/covariates.R, R/arm.R, R/arms.R,
%   R/dataset.R
\name{get_time_varying_covariates}
\alias{get_time_varying_covariates}
\alias{get_time_varying_covariates,covariates-method}
\alias{get_time_varying_covariates,arm-method}
\alias{get_time_varying_covariates,arms-method}
\alias{get_time_varying_covariates,dataset-method}
\title{Get all time-varying covariates.}
\usage{
get_time_varying_covariates(object)

\S4method{get_time_varying_covariates}{covariates}(object)

\S4method{get_time_varying_covariates}{arm}(object)

\S4method{get_time_varying_covariates}{arms}(object)

\S4method{get_time_varying_covariates}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
all time-varying covariates from object
}
\description{
Get all time-varying covariates.
}
