% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment.R, R/observations.R,
%   R/observations_set.R, R/arm.R, R/arms.R, R/events.R, R/dataset.R
\name{get_times}
\alias{get_times}
\alias{get_times,treatment-method}
\alias{get_times,observations-method}
\alias{get_times,observations_set-method}
\alias{get_times,arm-method}
\alias{get_times,arms-method}
\alias{get_times,events-method}
\alias{get_times,dataset-method}
\title{Get all distinct times for the specified object.}
\usage{
get_times(object, ...)

\S4method{get_times}{treatment}(object, unwrap = TRUE)

\S4method{get_times}{observations}(object, doseTimes = NULL)

\S4method{get_times}{observations_set}(object, doseTimes = NULL)

\S4method{get_times}{arm}(object)

\S4method{get_times}{arms}(object)

\S4method{get_times}{events}(object)

\S4method{get_times}{dataset}(object)
}
\arguments{
\item{object}{any object}

\item{...}{extra arguments like `doseTimes` in observations or `unwrap` in treatment}

\item{unwrap}{unwrap treatment before accessing the times, default value is TRUE}

\item{doseTimes}{times of the doses, only needed if a [DosingSchedule()] is referred to}
}
\value{
numeric vector with all unique times, sorted
}
\description{
Get all distinct times for the specified object.
}
