% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utilities.R
\name{getAvailableTimeUnits}
\alias{getAvailableTimeUnits}
\title{Return the list of available time units.}
\usage{
getAvailableTimeUnits()
}
\value{
character vector
}
\description{
Return the list of available time units.
}
