% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{getByCompartment}
\alias{getByCompartment}
\title{Get element by compartment number.}
\usage{
getByCompartment(object, compartment)
}
\arguments{
\item{object}{any object}

\item{compartment}{compartment number}
}
\value{
the element that has the right compartment
}
\description{
Get element by compartment number.
}
