% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCampsisOption}
\alias{getCampsisOption}
\title{Get Campsis option logic.}
\usage{
getCampsisOption(name, default)
}
\arguments{
\item{name}{option to search}

\item{default}{default value if option not found}
}
\value{
option value
}
\description{
Get Campsis option logic.
}
