% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCampsisOptions}
\alias{getCampsisOptions}
\title{Get the Campsis options (R options).}
\usage{
getCampsisOptions()
}
\value{
global options for Campsis
}
\description{
Get the Campsis options (R options).
}
\keyword{internal}
