% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{getColumn}
\alias{getColumn}
\title{Get data of given column unless if does not exist (return NULL in that case).}
\usage{
getColumn(.data, colname)
}
\arguments{
\item{.data}{data frame}

\item{colname}{column name}
}
\value{
a vector
}
\description{
Get data of given column unless if does not exist (return NULL in that case).
}
\keyword{internal}
