% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/occasion.R, R/treatment_iov.R
\name{getColumnName}
\alias{getColumnName}
\alias{getColumnName,occasion-method}
\alias{getColumnName,treatment_iov-method}
\title{Get column name.}
\usage{
getColumnName(x)

\S4method{getColumnName}{occasion}(x)

\S4method{getColumnName}{treatment_iov}(x)
}
\arguments{
\item{x}{any object}
}
\value{
column name
}
\description{
Get column name.
}
