% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{getCompartmentMapping}
\alias{getCompartmentMapping}
\title{Get a mapping table with all possibilities of compartment names and their indexes
knowing that compartment names can be provided as character or as integer.}
\usage{
getCompartmentMapping(compartments)
}
\arguments{
\item{compartments}{list of compartments}
}
\value{
a tibble with two columns: INDEX and NAME
}
\description{
Get a mapping table with all possibilities of compartment names and their indexes
knowing that compartment names can be provided as character or as integer.
}
\keyword{internal}
