% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{getCovariateNames}
\alias{getCovariateNames}
\title{Get all covariate names (fixed covariates + time-varying covariates).}
\usage{
getCovariateNames(object)
}
\arguments{
\item{object}{any object}
}
\value{
character vector
}
\description{
Get all covariate names (fixed covariates + time-varying covariates).
}
