% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/covariates.R, R/arm.R, R/arms.R,
%   R/dataset.R
\name{getCovariates}
\alias{getCovariates}
\alias{getCovariates,covariates-method}
\alias{getCovariates,arm-method}
\alias{getCovariates,arms-method}
\alias{getCovariates,dataset-method}
\title{Get all covariates (fixed / time-varying / event covariates).}
\usage{
getCovariates(object)

\S4method{getCovariates}{covariates}(object)

\S4method{getCovariates}{arm}(object)

\S4method{getCovariates}{arms}(object)

\S4method{getCovariates}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
all covariates from object
}
\description{
Get all covariates (fixed / time-varying / event covariates).
}
