% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_logic.R
\name{getDatasetMaxTime}
\alias{getDatasetMaxTime}
\title{Get dataset max time.}
\usage{
getDatasetMaxTime(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\value{
max time of dataset, whatever its form, 2-dimensional or structured
}
\description{
Get dataset max time.
}
\keyword{internal}
