% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/covariates.R, R/arm.R, R/arms.R,
%   R/dataset.R
\name{getEventCovariates}
\alias{getEventCovariates}
\alias{getEventCovariates,covariates-method}
\alias{getEventCovariates,arm-method}
\alias{getEventCovariates,arms-method}
\alias{getEventCovariates,dataset-method}
\title{Get all event-related covariates.}
\usage{
getEventCovariates(object)

\S4method{getEventCovariates}{covariates}(object)

\S4method{getEventCovariates}{arm}(object)

\S4method{getEventCovariates}{arms}(object)

\S4method{getEventCovariates}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
all event-related covariates from object
}
\description{
Get all event-related covariates.
}
