% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{getEventTimes}
\alias{getEventTimes}
\title{Get list of event times.}
\usage{
getEventTimes(events, maxTime)
}
\arguments{
\item{events}{events}

\item{maxTime}{simulation max time}
}
\value{
a list of numeric lists with names 'start' and 'end'
}
\description{
Get list of event times.
}
\keyword{internal}
