% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/covariates.R, R/arm.R, R/arms.R,
%   R/dataset.R
\name{getFixedCovariates}
\alias{getFixedCovariates}
\alias{getFixedCovariates,covariates-method}
\alias{getFixedCovariates,arm-method}
\alias{getFixedCovariates,arms-method}
\alias{getFixedCovariates,dataset-method}
\title{Get all fixed covariates.}
\usage{
getFixedCovariates(object)

\S4method{getFixedCovariates}{covariates}(object)

\S4method{getFixedCovariates}{arm}(object)

\S4method{getFixedCovariates}{arms}(object)

\S4method{getFixedCovariates}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
all fixed covariates from object
}
\description{
Get all fixed covariates.
}
