% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_setup.R
\name{getFurrrScheduling}
\alias{getFurrrScheduling}
\title{Get scheduling mode for furrr (see argument 'scheduling' available in furrr options).}
\usage{
getFurrrScheduling(parallel)
}
\arguments{
\item{parallel}{use parallel computing with furrr, logical value}
}
\value{
1 for parallel computing, 0 otherwise
}
\description{
Get scheduling mode for furrr (see argument 'scheduling' available in furrr options).
}
\keyword{internal}
