% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{getInitialConditions}
\alias{getInitialConditions}
\title{Get initial conditions at simulation start-up.}
\usage{
getInitialConditions(subdataset, iteration, cmtNames)
}
\arguments{
\item{subdataset}{subdataset to simulate}

\item{iteration}{current iteration}

\item{cmtNames}{compartment names}
}
\value{
named numeric vector with the new initial conditions
}
\description{
Get initial conditions at simulation start-up.
}
\keyword{internal}
