% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/arm.R, R/arms.R, R/dataset.R
\name{getIOVNames}
\alias{getIOVNames}
\alias{getIOVNames,arm-method}
\alias{getIOVNames,arms-method}
\alias{getIOVNames,dataset-method}
\title{Get IOV names.}
\usage{
getIOVNames(object)

\S4method{getIOVNames}{arm}(object)

\S4method{getIOVNames}{arms}(object)

\S4method{getIOVNames}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
character vector
}
\description{
Get IOV names.
}
