% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/arm.R, R/arms.R, R/dataset.R
\name{getIOVs}
\alias{getIOVs}
\alias{getIOVs,arm-method}
\alias{getIOVs,arms-method}
\alias{getIOVs,dataset-method}
\title{Get all IOV objects.}
\usage{
getIOVs(object)

\S4method{getIOVs}{arm}(object)

\S4method{getIOVs}{arms}(object)

\S4method{getIOVs}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
all IOV's from object
}
\description{
Get all IOV objects.
}
