% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_library.R
\name{getNONMEMModelTemplate}
\alias{getNONMEMModelTemplate}
\title{Get NONMEM model template.}
\usage{
getNONMEMModelTemplate(advan, trans)
}
\arguments{
\item{advan}{ADVAN number}

\item{trans}{TRANS number}
}
\value{
a PMX model ready to be used
}
\description{
Get NONMEM model template.
}
