% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{getOccasionNames}
\alias{getOccasionNames}
\title{Get all occasion names.}
\usage{
getOccasionNames(object)
}
\arguments{
\item{object}{any object}
}
\value{
character vector
}
\description{
Get all occasion names.
}
