% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/arm.R, R/arms.R, R/dataset.R
\name{getOccasions}
\alias{getOccasions}
\alias{getOccasions,arm-method}
\alias{getOccasions,arms-method}
\alias{getOccasions,dataset-method}
\title{Get all occasions.}
\usage{
getOccasions(object)

\S4method{getOccasions}{arm}(object)

\S4method{getOccasions}{arms}(object)

\S4method{getOccasions}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
all occasions from object
}
\description{
Get all occasions.
}
