% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{getRandomSeedValue}
\alias{getRandomSeedValue}
\title{Get random seed value.}
\usage{
getRandomSeedValue()
}
\value{
random seed value generated based on time
}
\description{
Get random seed value.
}
\keyword{internal}
