% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{getSeed}
\alias{getSeed}
\title{Get seed value.}
\usage{
getSeed(seed = NULL)
}
\arguments{
\item{seed}{user-input seed, NULL it not specified}
}
\value{
a seed value
}
\description{
Get seed value.
}
\keyword{internal}
