% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{getSeedForDatasetExport}
\alias{getSeedForDatasetExport}
\title{Get seed for dataset export.}
\usage{
getSeedForDatasetExport(seed, replicate, iterations)
}
\arguments{
\item{seed}{original seed}

\item{replicate}{the current replicate number}

\item{iterations}{total number of iterations}
}
\value{
the seed value used to export the dataset
}
\description{
Get seed for dataset export.
}
