% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{getSeedForIteration}
\alias{getSeedForIteration}
\title{Get seed for iteration.}
\usage{
getSeedForIteration(seed, replicate, iterations, iteration)
}
\arguments{
\item{seed}{original seed}

\item{replicate}{the current replicate number}

\item{iterations}{total number of iterations}
}
\value{
the seed value to be used for the given replicate number and iteration
}
\description{
Get seed for iteration.
}
