% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{getSeedForParametersSampling}
\alias{getSeedForParametersSampling}
\title{Get seed for parameter uncertainty sampling.}
\usage{
getSeedForParametersSampling(seed)
}
\arguments{
\item{seed}{original seed}
}
\value{
the seed value used to sample parameter uncertainty
}
\description{
Get seed for parameter uncertainty sampling.
}
