% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{getSeedForReplicate}
\alias{getSeedForReplicate}
\title{Get seed for replicate.}
\usage{
getSeedForReplicate(originalSeed, replicate)
}
\arguments{
\item{originalSeed}{original seed}

\item{replicate}{the replicate number}
}
\value{
a new seed value for this replicate
}
\description{
Get seed for replicate.
}
\keyword{internal}
