% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{getSimulationEngineType}
\alias{getSimulationEngineType}
\title{Get simulation engine type.}
\usage{
getSimulationEngineType(dest)
}
\arguments{
\item{dest}{destination engine, string form}
}
\value{
simulation engine type
}
\description{
Get simulation engine type.
}
\keyword{internal}
