% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{getSplittingConfiguration}
\alias{getSplittingConfiguration}
\title{Get splitting configuration for parallel export.}
\usage{
getSplittingConfiguration(dataset, hardware)
}
\arguments{
\item{dataset}{Campsis dataset to export}

\item{hardware}{hardware configuration}
}
\value{
splitting configuration list (if 'parallel_dataset' is enabled) or
 NA (if 'parallel_dataset' disabled or if the length of the dataset is less than the dataset export slice size)
}
\description{
Get splitting configuration for parallel export.
}
