% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/observations_set.R, R/arm.R,
%   R/arms.R, R/events.R, R/dataset.R
\name{getTimes}
\alias{getTimes}
\alias{getTimes,observations_set-method}
\alias{getTimes,arm-method}
\alias{getTimes,arms-method}
\alias{getTimes,events-method}
\alias{getTimes,dataset-method}
\title{Get all distinct times for the specified object.}
\usage{
getTimes(object)

\S4method{getTimes}{observations_set}(object)

\S4method{getTimes}{arm}(object)

\S4method{getTimes}{arms}(object)

\S4method{getTimes}{events}(object)

\S4method{getTimes}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
numeric vector with all unique times, sorted
}
\description{
Get all distinct times for the specified object.
}
