% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/arm.R, R/arms.R, R/dataset.R
\name{getTimeVaryingCovariateNames}
\alias{getTimeVaryingCovariateNames}
\alias{getTimeVaryingCovariateNames,arm-method}
\alias{getTimeVaryingCovariateNames,arms-method}
\alias{getTimeVaryingCovariateNames,dataset-method}
\title{Get all time-varying covariate names.}
\usage{
getTimeVaryingCovariateNames(object)

\S4method{getTimeVaryingCovariateNames}{arm}(object)

\S4method{getTimeVaryingCovariateNames}{arms}(object)

\S4method{getTimeVaryingCovariateNames}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
character vector
}
\description{
Get all time-varying covariate names.
}
