% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/covariates.R, R/arm.R, R/arms.R,
%   R/dataset.R
\name{getTimeVaryingCovariates}
\alias{getTimeVaryingCovariates}
\alias{getTimeVaryingCovariates,covariates-method}
\alias{getTimeVaryingCovariates,arm-method}
\alias{getTimeVaryingCovariates,arms-method}
\alias{getTimeVaryingCovariates,dataset-method}
\title{Get all time-varying covariates.}
\usage{
getTimeVaryingCovariates(object)

\S4method{getTimeVaryingCovariates}{covariates}(object)

\S4method{getTimeVaryingCovariates}{arm}(object)

\S4method{getTimeVaryingCovariates}{arms}(object)

\S4method{getTimeVaryingCovariates}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
all time-varying covariates from object
}
\description{
Get all time-varying covariates.
}
