% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{getTimeVaryingVariables}
\alias{getTimeVaryingVariables}
\title{Get all time-varying variables. These variables are likely to be influenced
by the NOCB mode chosen and by the 'nocbvars' vector.}
\usage{
getTimeVaryingVariables(object)
}
\arguments{
\item{object}{dataset}
}
\value{
character vector with all time-varying variables of the dataset
}
\description{
Get all time-varying variables. These variables are likely to be influenced
by the NOCB mode chosen and by the 'nocbvars' vector.
}
\keyword{internal}
