% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardware_settings.R
\docType{class}
\name{hardware_settings-class}
\alias{hardware_settings-class}
\title{Hardware settings class.}
\description{
Hardware settings class.
}
\section{Slots}{

\describe{
\item{\code{cpu}}{number of CPU's to use, default is 6}

\item{\code{parallel}}{enable parallel computing, logical value}

\item{\code{slices}}{number of subjects per simulated slice, default is 6}
}}

