% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{hasElement}
\alias{hasElement}
\title{Check if an element exists in list.}
\usage{
hasElement(object, x)
}
\arguments{
\item{object}{object containing a list}

\item{x}{element to check if exists}
}
\value{
logical value
}
\description{
Check if an element exists in list.
}
