% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{hasParameterDistribution}
\alias{hasParameterDistribution}
\title{Tell if the dataset contains at least one parameter distribution (in its treatment characteristics).
This is useful information to know as we therefore need the PMX model to know the
corresponding values (THETA and OMEGA) for this parameter.}
\usage{
hasParameterDistribution(object)
}
\arguments{
\item{object}{generic object}
}
\value{
logical value
}
\description{
Tell if the dataset contains at least one parameter distribution (in its treatment characteristics).
This is useful information to know as we therefore need the PMX model to know the
corresponding values (THETA and OMEGA) for this parameter.
}
