% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utilities.R
\name{hours}
\alias{hours}
\title{Convert hours to hours (do nothing).}
\usage{
hours(x)
}
\arguments{
\item{x}{numeric vector in hours}
}
\value{
numeric vector in hours
}
\description{
Convert hours to hours (do nothing).
}
