% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{import_campsismod_to_namespace}
\alias{import_campsismod_to_namespace}
\title{Import the whole campsismod package into NAMESPACE when parsed by 'roxygen'.}
\usage{
import_campsismod_to_namespace()
}
\value{
always TRUE
}
\description{
Import the whole campsismod package into NAMESPACE when parsed by 'roxygen'.
}
\keyword{internal}
