% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{importCampsismodToNamespace}
\alias{importCampsismodToNamespace}
\title{Import the whole campsismod package into NAMESPACE when parsed by roxygen.}
\usage{
importCampsismodToNamespace()
}
\description{
Import the whole campsismod package into NAMESPACE when parsed by roxygen.
}
\keyword{internal}
