% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{indexOf}
\alias{indexOf}
\title{Get the index of an element in list.}
\usage{
indexOf(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{element to know the index}
}
\value{
index of this element
}
\description{
Get the index of an element in list.
}
