% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_entry.R
\docType{class}
\name{infusion-class}
\alias{infusion-class}
\title{Infusion class.}
\description{
Infusion class.
}
\section{Slots}{

\describe{
\item{\code{duration}}{infusion duration, distribution list}

\item{\code{rate}}{infusion rate, distribution list}
}}

