% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_settings.R
\docType{class}
\name{internal_settings-class}
\alias{internal_settings-class}
\title{Internal settings class (transient object from the simulation settings).}
\description{
Internal settings class (transient object from the simulation settings).
}
\section{Slots}{

\describe{
\item{\code{dataset_summary}}{dataset summary}

\item{\code{progress}}{simulation progress}

\item{\code{iterations}}{list of event iterations}
}}

