% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_iov.R
\docType{class}
\name{iov-class}
\alias{iov-class}
\title{Treatment IOV class.}
\description{
Treatment IOV class.
}
\section{Slots}{

\describe{
\item{\code{colname}}{name of the column that will be output in dataset}

\item{\code{distribution}}{distribution}

\item{\code{dose_numbers}}{associated dose numbers, integer vector, same length as values}

\item{\code{omega_ref}}{name of the OMEGA (e.g. 'IOV' without the 'OMEGA_')}
}}

