% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_interface.R
\name{jsonToCampsisSettings}
\alias{jsonToCampsisSettings}
\title{JSON to Campsis dataset.}
\usage{
jsonToCampsisSettings(object, json)
}
\arguments{
\item{object}{empty dataset}

\item{json}{json element}
}
\value{
Campsis dataset
}
\description{
JSON to Campsis dataset.
}
\keyword{internal}
