% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{left_join_iiv}
\alias{left_join_iiv}
\title{Left-join IIV matrix.}
\usage{
left_join_iiv(table, iiv)
}
\arguments{
\item{table}{dataset, tabular form}

\item{iiv}{IIV matrix}
}
\value{
updated table with IIV matrix
}
\description{
Left-join IIV matrix.
}
\keyword{internal}
