% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.R
\name{length,arm-method}
\alias{length,arm-method}
\title{Return the number of subjects contained in this arm.}
\usage{
\S4method{length}{arm}(x)
}
\arguments{
\item{x}{arm}
}
\value{
a number
}
\description{
Return the number of subjects contained in this arm.
}
