% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_schedule.R
\name{length,cyclic_schedule-method}
\alias{length,cyclic_schedule-method}
\title{Return the number of repetition cycles.}
\usage{
\S4method{length}{cyclic_schedule}(x)
}
\arguments{
\item{x}{schedule object}
}
\value{
a number
}
\description{
Return the number of repetition cycles.
}
