% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{length,dataset-method}
\alias{length,dataset-method}
\title{Return the number of subjects contained in this dataset.}
\usage{
\S4method{length}{dataset}(x)
}
\arguments{
\item{x}{dataset}
}
\value{
a number
}
\description{
Return the number of subjects contained in this dataset.
}
