% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment_entry.R,
%   R/observations.R, R/dataset.R
\name{loadFromJSON}
\alias{loadFromJSON}
\alias{loadFromJSON,bolus_wrapper-method}
\alias{loadFromJSON,observations-method}
\alias{loadFromJSON,dataset-method}
\title{Fill-in S4 object from the JSON content.}
\usage{
loadFromJSON(object, json)

\S4method{loadFromJSON}{bolus_wrapper}(object, json)

\S4method{loadFromJSON}{observations}(object, json)

\S4method{loadFromJSON}{dataset}(object, json)
}
\arguments{
\item{object}{pre-initiated S4 object}

\item{json}{JSON (usually a list)}
}
\value{
the loaded S4 object
}
\description{
Fill-in S4 object from the JSON content.
}
