% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{LogNormalDistribution}
\alias{LogNormalDistribution}
\title{Create a log normal distribution.}
\usage{
LogNormalDistribution(meanlog, sdlog)
}
\arguments{
\item{meanlog}{mean value of distribution in log domain}

\item{sdlog}{standard deviation of distribution in log domain}
}
\value{
a log normal distribution
}
\description{
Create a log normal distribution.
}
