% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun_core.R
\name{make_vpc_summary}
\alias{make_vpc_summary}
\title{Make the VPC summary. Input data frame must contain the following columns:
- replicate: replicate number
- low: low percentile value in replicate (and in scenario if present)
- med: median value in replicate (and in scenario if present)
- up: up percentile value in replicate (and in scenario if present)
- any scenario column}
\usage{
make_vpc_summary(x, strata = NULL, level = 0.9)
}
\arguments{
\item{x}{data frame, with metric}

\item{strata}{named vector with the strata to use, default is c(SCENARIO="all", ARM="all").
Only columns that are actually present in \code{x} are used.}

\item{level}{PI level, default is 0.9 (90\% PI)}
}
\value{
VPC summary with columns TIME, stratification variables and all combinations of
low, med, up (i.e. low_low, low_med, low_up, etc.)
}
\description{
Make the VPC summary. Input data frame must contain the following columns:
- replicate: replicate number
- low: low percentile value in replicate (and in scenario if present)
- med: median value in replicate (and in scenario if present)
- up: up percentile value in replicate (and in scenario if present)
- any scenario column
}
\keyword{internal}
