% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{methodsGlobalFunctions}
\alias{methodsGlobalFunctions}
\title{A empty method only used to declare some of the global function definitions
contained in the package 'methods', which are used intensively.}
\usage{
methodsGlobalFunctions()
}
\description{
A empty method only used to declare some of the global function definitions
contained in the package 'methods', which are used intensively.
}
\keyword{internal}
