% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utilities.R
\name{minutes}
\alias{minutes}
\title{Convert minutes to hours.}
\usage{
minutes(x)
}
\arguments{
\item{x}{numeric vector in minutes}
}
\value{
numeric vector in hours
}
\description{
Convert minutes to hours.
}
