% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_library}
\alias{model_library}
\title{PMXmod model library.}
\format{
A list with all the models:
\describe{
}
}
\source{
\url{https://www.iconplc.com/innovation/nonmem/}
}
\usage{
model_library
}
\description{
A list containing model templates.
}
\keyword{datasets}
