% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utilities.R
\name{months}
\alias{months}
\title{Convert pharma months (1 month = 4 weeks) to hours.}
\usage{
months(x)
}
\arguments{
\item{x}{numeric vector in months}
}
\value{
numeric vector in hours
}
\description{
Convert pharma months (1 month = 4 weeks) to hours.
}
