% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun.R
\docType{class}
\name{nca_table_outfun-class}
\alias{nca_table_outfun-class}
\title{NCA table output function class.}
\description{
NCA table output function class.
}
\section{Slots}{

\describe{
\item{\code{table}}{Campsisnca table object}

\item{\code{export_type}}{type of export, 'summary', 'summary_wide', 'summary_pretty', 'individual' or 'individual_wide'}
}}

