% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun.R
\name{NCATableOutfun}
\alias{NCATableOutfun}
\title{Create a NCA table output function}
\usage{
NCATableOutfun(table, export_type = "summary", name = "default_nca_table")
}
\arguments{
\item{table}{NCA table from campsisnca, object, path to JSON, or JSON element}

\item{export_type}{type of export, 'summary', 'summary_wide', 'summary_pretty', 'individual' or 'individual_wide'}

\item{name}{name of the output function. Default is 'default_nca_table'.}
}
\value{
a stats_outfun object
}
\description{
Create a NCA table output function
}
