% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{NHANES database (demographics and body measure data combined, from 2017-2018).}
\format{
data frame
\describe{
   \item{ID}{Original identifier}
   \item{SEX}{Sex: 1 for males, 2 for females}
   \item{AGE}{Age in years}
   \item{BW}{Body weight in kg}
   \item{BMI}{Body mass index}
   \item{HT}{Height in cm}
}
}
\source{
\url{https://wwwn.cdc.gov/Nchs/Nhanes/2017-2018/DEMO_J.XPT}

\url{https://wwwn.cdc.gov/Nchs/Nhanes/2017-2018/BMX_J.XPT}
}
\usage{
nhanes
}
\description{
NHANES database (demographics and body measure data combined, from 2017-2018).
}
\keyword{datasets}
