% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocb_settings.R
\name{NOCB}
\alias{NOCB}
\title{Create NOCB settings.}
\usage{
NOCB(enable = NULL, variables = character(0))
}
\arguments{
\item{enable}{enable/disable next-observation carried backward mode (NOCB), default value is TRUE for mrgsolve, FALSE for RxODE}

\item{variables}{variable names subject to NOCB behavior (see vignette for more info)}
}
\value{
NOCB settings
}
\description{
Create NOCB settings.
}
