% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocb_settings.R
\docType{class}
\name{nocb_settings-class}
\alias{nocb_settings-class}
\title{NOCB settings class.}
\description{
NOCB settings class.
}
\section{Slots}{

\describe{
\item{\code{enable}}{enable/disable next-observation carried backward mode (NOCB), default value is TRUE for mrgsolve, FALSE for RxODE}

\item{\code{variables}}{variable names subject to NOCB behavior (see vignette for more info)}
}}

