% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{Observation}
\alias{Observation}
\title{Create an observation.}
\usage{
Observation(time, compartment = NA)
}
\arguments{
\item{time}{observation time, numeric}

\item{compartment}{compartment index, integer}
}
\value{
an observation
}
\description{
Create an observation.
}
