% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\docType{class}
\name{observations-class}
\alias{observations-class}
\title{Observations class.}
\description{
Observations class.
}
\section{Slots}{

\describe{
\item{\code{times}}{observation times, numeric vector}

\item{\code{compartment}}{compartment index (integer) or name (character)}

\item{\code{dv}}{observed values, numeric vector (FOR EXTERNAL USE)}
}}

