% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{obsOnly}
\alias{obsOnly}
\title{Filter CAMPSIS output on observation rows.}
\usage{
obsOnly(x)
}
\arguments{
\item{x}{data frame, CAMPSIS output}
}
\value{
a data frame with the observation rows
}
\description{
Filter CAMPSIS output on observation rows.
}
