% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occasion.R
\name{Occasion}
\alias{Occasion}
\title{Define a new occasion. Occasions are defined by mapping occasion values to dose numbers.
A new column will automatically be created in the exported dataset.}
\usage{
Occasion(colname, values, dose_numbers)
}
\arguments{
\item{colname}{name of the column that will be output in dataset}

\item{values}{the occasion numbers, any integer vector}

\item{dose_numbers}{the related dose numbers, any integer vector of same length as 'values'}
}
\value{
occasion object
}
\description{
Define a new occasion. Occasions are defined by mapping occasion values to dose numbers.
A new column will automatically be created in the exported dataset.
}
