% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{on_cran}
\alias{on_cran}
\title{Check if the current session is on CRAN. The objective is to potentially suppress
long tasks to be run on CRAN (long tests or vignettes).}
\usage{
on_cran()
}
\value{
logical value TRUE/FALSE
}
\description{
Check if the current session is on CRAN. The objective is to potentially suppress
long tasks to be run on CRAN (long tests or vignettes).
}
\keyword{internal}
