% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{onCI}
\alias{onCI}
\title{Check if the current session is on CI (e.g. GitHub actions).}
\usage{
onCI()
}
\value{
logical value TRUE/FALSE
}
\description{
Check if the current session is on CI (e.g. GitHub actions).
}
\keyword{internal}
