% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun.R
\docType{class}
\name{output_function-class}
\alias{output_function-class}
\title{Output function class.}
\description{
Output function class.
}
\section{Slots}{

\describe{
\item{\code{fun}}{function or purrr-style lambda formula}

\item{\code{level}}{either 'scenario' or 'replicate'. Default is 'scenario'.}
}}

