% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{ParameterDistribution}
\alias{ParameterDistribution}
\title{Create a parameter distribution. The resulting distribution is a
log-normal distribution, with meanlog=log(THETA) and sdlog=sqrt(OMEGA).}
\usage{
ParameterDistribution(model, theta, omega = NULL)
}
\arguments{
\item{model}{model}

\item{theta}{corresponding THETA name, character}

\item{omega}{corresponding OMEGA name, character, NULL if not defined}
}
\value{
a parameter distribution
}
\description{
Create a parameter distribution. The resulting distribution is a
log-normal distribution, with meanlog=log(THETA) and sdlog=sqrt(OMEGA).
}
