% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{pi_campsis_tbl-class}
\alias{pi_campsis_tbl-class}
\alias{pi_campsis_tbl}
\title{Prediction Interval (PI) Campsis table class}
\description{
Prediction Interval (PI) Campsis table class
}
