% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun.R
\docType{class}
\name{pi_outfun-class}
\alias{pi_outfun-class}
\title{Prediction interval output function class.}
\description{
Prediction interval output function class.
}
\section{Slots}{

\describe{
\item{\code{variable}}{variable(s) used to compute the prediction interval, character vector}

\item{\code{strata}}{named vector with the strata to use}

\item{\code{level}}{PI level, default is 0.9 (90\% PI)}

\item{\code{gather}}{FALSE: med, low & up columns, TRUE: metric column}
}}

