% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocess_arm_column}
\alias{preprocess_arm_column}
\title{Preprocess ARM column. Add ARM equation in model automatically.}
\usage{
preprocess_arm_column(dataset, model)
}
\arguments{
\item{dataset}{current dataset, data frame form}

\item{model}{model}
}
\value{
updated model
}
\description{
Preprocess ARM column. Add ARM equation in model automatically.
}
\keyword{internal}
