% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocess_dest}
\alias{preprocess_dest}
\title{Pre-process destination engine. Throw an error message if the destination
engine is not installed.}
\usage{
preprocess_dest(dest)
}
\arguments{
\item{dest}{destination engine (from simulate argument, default settings or S4 simulation engine)}
}
\value{
'rxode2' or 'mrgsolve'
}
\description{
Pre-process destination engine. Throw an error message if the destination
engine is not installed.
}
\keyword{internal}
