% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocess_dosing}
\alias{preprocess_dosing}
\title{Preprocess 'dosing' argument.}
\usage{
preprocess_dosing(dosing)
}
\arguments{
\item{dosing}{dosing argument, logical value}
}
\value{
user value, if not specified, return FALSE (observations only)
}
\description{
Preprocess 'dosing' argument.
}
\keyword{internal}
