% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocess_tablefun}
\alias{preprocess_tablefun}
\title{Pre-process tablefun argument.}
\usage{
preprocess_tablefun(fun)
}
\arguments{
\item{fun}{function or lambda formula}
}
\value{
a function in any case
}
\description{
Pre-process tablefun argument.
}
\keyword{internal}
