% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocessArmColumn}
\alias{preprocessArmColumn}
\title{Preprocess ARM column. Add ARM equation in model automatically.}
\usage{
preprocessArmColumn(dataset, model)
}
\arguments{
\item{dataset}{current dataset, data frame form}

\item{model}{model}
}
\value{
updated model
}
\description{
Preprocess ARM column. Add ARM equation in model automatically.
}
\keyword{internal}
