% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocessDest}
\alias{preprocessDest}
\title{Pre-process destination engine. Throw an error message if the destination 
engine is not installed.}
\usage{
preprocessDest(dest)
}
\arguments{
\item{dest}{destination engine}
}
\value{
'RxODE', 'mrgsolve'
}
\description{
Pre-process destination engine. Throw an error message if the destination 
engine is not installed.
}
\keyword{internal}
