% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{preprocessFunction}
\alias{preprocessFunction}
\title{Pre-process function argument.}
\usage{
preprocessFunction(fun, name)
}
\arguments{
\item{fun}{function or lambda formula}

\item{name}{function name}
}
\value{
a function in any case
}
\description{
Pre-process function argument.
}
\keyword{internal}
