% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocessNocb}
\alias{preprocessNocb}
\title{Preprocess 'nocb' argument.}
\usage{
preprocessNocb(nocb)
}
\arguments{
\item{nocb}{nocb argument, logical value}
}
\value{
same value
}
\description{
Preprocess 'nocb' argument.
}
\keyword{internal}
