% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{preprocessNocbvars}
\alias{preprocessNocbvars}
\title{Preprocess 'nocbvars' argument.}
\usage{
preprocessNocbvars(nocbvars)
}
\arguments{
\item{nocbvars}{nocbvars argument, character vector}
}
\description{
Preprocess 'nocbvars' argument.
}
\keyword{internal}
