% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{preprocessOutput}
\alias{preprocessOutput}
\title{Preprocess 'output' argument. Output is either the columns to keep from the 
returned dataframe or the function to apply on this dataframe.}
\usage{
preprocessOutput(output)
}
\arguments{
\item{output}{character vector or function}
}
\value{
output
}
\description{
Preprocess 'output' argument. Output is either the columns to keep from the 
returned dataframe or the function to apply on this dataframe.
}
