% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{preprocessOutvars}
\alias{preprocessOutvars}
\title{Preprocess 'outvars' argument. Outvars is a character vector that contains 
all the columns to keep in the RxODE/mrgsolve output.}
\usage{
preprocessOutvars(outvars, model, dest)
}
\arguments{
\item{outvars}{character vector or function}
}
\value{
outvars
}
\description{
Preprocess 'outvars' argument. Outvars is a character vector that contains 
all the columns to keep in the RxODE/mrgsolve output.
}
