% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{preprocessReplicates}
\alias{preprocessReplicates}
\title{Preprocess 'replicates' argument.}
\usage{
preprocessReplicates(replicates)
}
\arguments{
\item{replicates}{number of replicates}
}
\value{
same number
}
\description{
Preprocess 'replicates' argument.
}
\keyword{internal}
