% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocessScenarios}
\alias{preprocessScenarios}
\title{Pre-process scenarios.}
\usage{
preprocessScenarios(scenarios)
}
\arguments{
\item{scenarios}{scenarios}
}
\description{
Pre-process scenarios.
}
\keyword{internal}
