% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocessSettings}
\alias{preprocessSettings}
\title{Preprocess the simulation settings.}
\usage{
preprocessSettings(settings, dest)
}
\arguments{
\item{settings}{simulation settings}

\item{dest}{destination engine}
}
\value{
updated simulation settings
}
\description{
Preprocess the simulation settings.
}
\keyword{internal}
