% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{preprocessSimulateArguments}
\alias{preprocessSimulateArguments}
\title{Preprocess arguments of the simulate method.}
\usage{
preprocessSimulateArguments(model, dataset, dest, slices, ...)
}
\arguments{
\item{model}{PMX model}

\item{dataset}{dataset, data.frame form}

\item{dest}{destination engine}

\item{slices}{number of subjects per slice}
}
\value{
a simulation configuration
}
\description{
Preprocess arguments of the simulate method.
}
