% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{preprocessSlices}
\alias{preprocessSlices}
\title{Preprocess 'slices' argument.}
\usage{
preprocessSlices(slices, maxID)
}
\arguments{
\item{slices}{slices argument corresponding to the number of subjects simulated at once}
}
\value{
slices if not NULL, otherwise total number of subjects
}
\description{
Preprocess 'slices' argument.
}
\keyword{internal}
