% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{preprocessTSLDAndTDOSColumn}
\alias{preprocessTSLDAndTDOSColumn}
\title{Preprocess TSLD and TDOS columns according to given dataset configuration.}
\usage{
preprocessTSLDAndTDOSColumn(table, config)
}
\arguments{
\item{table}{current table}

\item{config}{dataset config}
}
\value{
updated table
}
\description{
Preprocess TSLD and TDOS columns according to given dataset configuration.
}
\keyword{internal}
