% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{process_arm_labels}
\alias{process_arm_labels}
\title{Process arm labels. Arm identifiers in ARM column are replaced by arm labels
as soon as one arm label is provided.}
\usage{
process_arm_labels(campsis, arms)
}
\arguments{
\item{campsis}{Campsis output}

\item{arms}{all treatment arms}
}
\value{
updated Campsis output with arm labels instead of arm identifiers
}
\description{
Process arm labels. Arm identifiers in ARM column are replaced by arm labels
as soon as one arm label is provided.
}
\keyword{internal}
