% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{processAllTimeColumns}
\alias{processAllTimeColumns}
\title{Process time-related columns according to given dataset configuration.}
\usage{
processAllTimeColumns(table, config)
}
\arguments{
\item{table}{current table}

\item{config}{dataset config}
}
\value{
updated table
}
\description{
Process time-related columns according to given dataset configuration.
}
\keyword{internal}
