% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{processArmLabels}
\alias{processArmLabels}
\title{Process arm labels. Arm identifiers in ARM column are replaced by arm labels
as soon as one arm label is provided.}
\usage{
processArmLabels(campsis, arms)
}
\arguments{
\item{campsis}{CAMPSIS output}

\item{arms}{all treatment arms}
}
\value{
updated CAMPSIS output with arm labels instead of arm identifiers
}
\description{
Process arm labels. Arm identifiers in ARM column are replaced by arm labels
as soon as one arm label is provided.
}
\keyword{internal}
