% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{processAsCovariate}
\alias{processAsCovariate}
\title{Process distribution-derived objects as covariates.}
\usage{
processAsCovariate(list)
}
\arguments{
\item{list}{list of distribution-derived objects}
}
\value{
a list of covariates that can be easily processed
}
\description{
Process distribution-derived objects as covariates.
}
\keyword{internal}
