% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{processBioavailabilities}
\alias{processBioavailabilities}
\title{Process bioavailabilities.}
\usage{
processBioavailabilities(table, characteristics)
}
\arguments{
\item{table}{current dataset, data frame form}

\item{characteristics}{all treatment characteristics}
}
\value{
updated table (AMT adapted)
}
\description{
Process bioavailabilities.
}
