% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{processCharacteristicIOV}
\alias{processCharacteristicIOV}
\title{Process characteristic IOV.}
\usage{
processCharacteristicIOV(table, characteristic)
}
\arguments{
\item{table}{current dataset, data frame form}

\item{characteristics}{all treatment characteristics}
}
\value{
updated table (AMT adapted)
}
\description{
Process characteristic IOV.
}
