% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{processDropOthers}
\alias{processDropOthers}
\title{Process 'DROP_OTHERS'.}
\usage{
processDropOthers(x, outvars = character(0), dropOthers)
}
\arguments{
\item{x}{the current data frame}

\item{outvars}{variables to keep}

\item{dropOthers}{logical value}
}
\value{
processed data frame
}
\description{
Process 'DROP_OTHERS'.
}
\keyword{internal}
