% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{processInfusions}
\alias{processInfusions}
\title{Process infusions.}
\usage{
processInfusions(table, characteristics)
}
\arguments{
\item{table}{current dataset, data frame form}

\item{characteristics}{all treatment characteristics}
}
\value{
updated table with RATE column
}
\description{
Process infusions.
}
