% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{processLagTimes}
\alias{processLagTimes}
\title{Process lag times.}
\usage{
processLagTimes(table, characteristics)
}
\arguments{
\item{table}{current dataset, data frame form}

\item{characteristics}{all treatment characteristics}
}
\value{
updated table with time of doses updated
}
\description{
Process lag times.
}
