% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{processOutput}
\alias{processOutput}
\title{Process the data frame returned by mrgsolve or RxODE.}
\usage{
processOutput(x, outfun = NULL)
}
\arguments{
\item{x}{the current data frame}

\item{outfun}{processing function}
}
\value{
processed data frame
}
\description{
Process the data frame returned by mrgsolve or RxODE.
}
