% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{processReturnedDataframe}
\alias{processReturnedDataframe}
\title{Process the data frame returned by mrgsolve or RxODE.}
\usage{
processReturnedDataframe(x, output)
}
\arguments{
\item{x}{the current data frame}

\item{output}{character vector or function}
}
\value{
processed data frame
}
\description{
Process the data frame returned by mrgsolve or RxODE.
}
