% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{processSimulateArguments}
\alias{processSimulateArguments}
\title{Preprocess arguments of the simulate method.}
\usage{
processSimulateArguments(model, dataset, dest, outvars, ...)
}
\arguments{
\item{model}{PMX model}

\item{dataset}{dataset, data.frame form}

\item{dest}{destination engine}

\item{outvars}{outvars}

\item{...}{all other arguments}
}
\value{
a simulation configuration
}
\description{
Preprocess arguments of the simulate method.
}
\keyword{internal}
