% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{processTable}
\alias{processTable}
\title{Process exported table.}
\usage{
processTable(x, tablefun = NULL)
}
\arguments{
\item{x}{the current data set, table form}

\item{tablefun}{function to apply}
}
\value{
processed data set, table form
}
\description{
Process exported table.
}
