% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{processTSLDAndTDOSColumn}
\alias{processTSLDAndTDOSColumn}
\title{Process TSLD and TDOS columns according to given dataset configuration.}
\usage{
processTSLDAndTDOSColumn(table, config)
}
\arguments{
\item{table}{current table}

\item{config}{dataset config}
}
\value{
updated table
}
\description{
Process TSLD and TDOS columns according to given dataset configuration.
}
\keyword{internal}
