% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_settings.R
\name{Progress}
\alias{Progress}
\title{Create progress settings.}
\usage{
Progress(tick_slice = TRUE)
}
\arguments{
\item{tick_slice}{tick() is called after each simulated slice, default is TRUE.
In some cases, when the number of subjects per slice is low, it may be useful disable this flag,
to improve performance issues.}
}
\value{
progress settings
}
\description{
Create progress settings.
}
