% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_settings.R
\docType{class}
\name{progress_settings-class}
\alias{progress_settings-class}
\title{Progress settings class.}
\description{
Progress settings class.
}
\section{Slots}{

\describe{
\item{\code{tick_slice}}{tick() is called after each simulated slice, default is TRUE.
In some cases, when the number of subjects per slice is low, it may be useful disable this flag,
to improve performance issues.}
}}

