% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_column_levels}
\alias{remove_column_levels}
\title{Strip Factor Class from Columns}
\usage{
remove_column_levels(x, cols)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{cols}{A character vector of column names to convert.}
}
\value{
A data frame with character columns.
}
\description{
Converts target columns from factors into standard character vectors.
}
