% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{remove_initial_conditions}
\alias{remove_initial_conditions}
\title{Remove initial conditions.}
\usage{
remove_initial_conditions(model)
}
\arguments{
\item{model}{Campsis model}
}
\value{
same model without initial conditions
}
\description{
Remove initial conditions.
}
\keyword{internal}
