% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{removeInitialConditions}
\alias{removeInitialConditions}
\title{Remove initial conditions.}
\usage{
removeInitialConditions(model)
}
\arguments{
\item{model}{PMX model}
}
\value{
same model without initial conditions
}
\description{
Remove initial conditions.
}
\keyword{internal}
