% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_schedule.R
\docType{class}
\name{repeat_at_schedule-class}
\alias{repeat_at_schedule-class}
\title{'Repeat at' schedule class.}
\description{
'Repeat at' schedule class.
}
\section{Slots}{

\describe{
\item{\code{times}}{times at which the event is repeated, numeric vector}
}}

