% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/repeated_schedule.R
\name{repeat_schedule}
\alias{repeat_schedule}
\alias{repeat_schedule,numeric,cyclic_schedule-method}
\alias{repeat_schedule,numeric,repeat_at_schedule-method}
\alias{repeat_schedule,numeric,undefined_schedule-method}
\title{Repeat schedule.}
\usage{
repeat_schedule(x, schedule)

\S4method{repeat_schedule}{numeric,cyclic_schedule}(x, schedule)

\S4method{repeat_schedule}{numeric,repeat_at_schedule}(x, schedule)

\S4method{repeat_schedule}{numeric,undefined_schedule}(x, schedule)
}
\arguments{
\item{x}{object to repeat the schedule}

\item{schedule}{initial times vector}
}
\value{
resulting times vector
}
\description{
Repeat schedule.
}
