% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_schedule.R
\name{RepeatAtSchedule}
\alias{RepeatAtSchedule}
\title{'Repeat at' schedule constructor. Note that the time 0 for the base pattern
 will be added by default if not provided.}
\usage{
RepeatAtSchedule(times)
}
\arguments{
\item{times}{times at which the original schedule must be repeated, numeric vector}
}
\value{
a 'repeat-at' schedule
}
\description{
'Repeat at' schedule constructor. Note that the time 0 for the base pattern
 will be added by default if not provided.
}
