% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_schedule.R
\docType{class}
\name{repeated_schedule-class}
\alias{repeated_schedule-class}
\title{Repeated schedule class. See this class as an interface.}
\description{
Repeated schedule class. See this class as an interface.
}
