% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/repeated_schedule.R
\name{repeatSchedule}
\alias{repeatSchedule}
\alias{repeatSchedule,numeric,cyclic_schedule-method}
\alias{repeatSchedule,numeric,repeat_at_schedule-method}
\alias{repeatSchedule,numeric,undefined_schedule-method}
\title{Repeat schedule.}
\usage{
repeatSchedule(x, schedule)

\S4method{repeatSchedule}{numeric,cyclic_schedule}(x, schedule)

\S4method{repeatSchedule}{numeric,repeat_at_schedule}(x, schedule)

\S4method{repeatSchedule}{numeric,undefined_schedule}(x, schedule)
}
\arguments{
\item{x}{object to repeat the schedule}

\item{schedule}{initial times vector}
}
\value{
resulting times vector
}
\description{
Repeat schedule.
}
