% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{replace}
\alias{replace}
\title{Replace element by another in list.}
\usage{
replace(object, x)
}
\arguments{
\item{object}{list object}

\item{x}{element to replace}
}
\value{
list object
}
\description{
Replace element by another in list.
}
