% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_action.R
\name{ReplaceAction}
\alias{ReplaceAction}
\title{Create a replace action.}
\usage{
ReplaceAction(object)
}
\arguments{
\item{object}{replacement object}
}
\value{
a replace action
}
\description{
Create a replace action.
}
