% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_uncertainty.R
\name{ReplicatedCampsisModel}
\alias{ReplicatedCampsisModel}
\title{Replicate a Campsis model for a study trial.
The replicated Campsis model must be seen as a list of Campsis models,
each one representing a trial, with its set of parameters sampled from the
original variance-covariance matrix.}
\usage{
ReplicatedCampsisModel(model, replicates)
}
\arguments{
\item{model}{original Campsis model to replicate}

\item{replicates}{number of replicates, integer}
}
\value{
a replicated Campsis model
}
\description{
Replicate a Campsis model for a study trial.
The replicated Campsis model must be seen as a list of Campsis models,
each one representing a trial, with its set of parameters sampled from the
original variance-covariance matrix.
}
