% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{sample}
\alias{sample}
\title{Sample generic object.}
\usage{
sample(object, n, ...)
}
\arguments{
\item{object}{generic object}

\item{n}{number of samples required}

\item{...}{extra arguments}
}
\value{
sampling result
}
\description{
Sample generic object.
}
