% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{sampleDistributionAsTibble}
\alias{sampleDistributionAsTibble}
\title{Sample a distribution and return a tibble.}
\usage{
sampleDistributionAsTibble(distribution, n, colname)
}
\arguments{
\item{distribution}{any distribution}

\item{n}{number of desired samples}

\item{colname}{name of the unique column in tibble}
}
\value{
a tibble of n rows and 1 column
}
\description{
Sample a distribution and return a tibble.
}
\keyword{internal}
