% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{sampleTimeVaryingCovariates}
\alias{sampleTimeVaryingCovariates}
\title{Sample time-varying covariates.}
\usage{
sampleTimeVaryingCovariates(object, armID, needsDV)
}
\arguments{
\item{object}{time-varying covariates, data.frame form}

\item{armID}{treatment arm ID}

\item{needsDV}{append extra column DV, logical value}
}
\value{
a data.frame
}
\description{
Sample time-varying covariates.
}
\keyword{internal}
