% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter plot (S3 generic).}
\usage{
scatter_plot(x, ...)
}
\arguments{
\item{x}{a Campsis output object}

\item{...}{additional arguments passed to the method}
}
\value{
a ggplot object
}
\description{
Scatter plot (S3 generic).
}
