% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\docType{class}
\name{scenario-class}
\alias{scenario-class}
\title{Scenario class.}
\description{
Scenario class.
}
\section{Slots}{

\describe{
\item{\code{name}}{scenario name, single character string}

\item{\code{model}}{either a Campsis model, a function or lambda-style formula}

\item{\code{dataset}}{either a Campsis dataset, a function or lambda-style formula}

\item{\code{actions}}{list of actions to apply}

\item{\code{enabled}}{is the scenario enabled in the simulation, logical value}
}}

