% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{Scenario}
\alias{Scenario}
\title{Create an scenario.}
\usage{
Scenario(name = NULL, model = NULL, dataset = NULL)
}
\value{
a new scenario
}
\description{
Create an scenario.
}
\section{Slots}{

\describe{
\item{\code{name}}{scenario name, single character string}

\item{\code{model}}{either a CAMPSIS model, a function or lambda-style formula}

\item{\code{dataset}}{either a CAMPSIS dataset, a function or lambda-style formula}
}}

