% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utilities.R
\name{seconds}
\alias{seconds}
\title{Convert seconds to hours.}
\usage{
seconds(x)
}
\arguments{
\item{x}{numeric vector in seconds}
}
\value{
numeric vector in hours
}
\description{
Convert seconds to hours.
}
