% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/arm.R, R/dataset.R
\name{set_subjects}
\alias{set_subjects}
\alias{set_subjects,arm,integer-method}
\alias{set_subjects,dataset,integer-method}
\title{Set the number of subjects.}
\usage{
set_subjects(object, x)

\S4method{set_subjects}{arm,integer}(object, x)

\S4method{set_subjects}{dataset,integer}(object, x)
}
\arguments{
\item{object}{any object}

\item{x}{the new number of subjects}
}
\value{
the updated object
}
\description{
Set the number of subjects.
}
