% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{setLabel}
\alias{setLabel}
\title{Set the label.}
\usage{
setLabel(object, x)
}
\arguments{
\item{object}{any object that has a label}

\item{x}{the new label}
}
\value{
the updated object
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`setLabel()` is deprecated in favor of `set_label()`.
}
\keyword{internal}
