% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{setSeed}
\alias{setSeed}
\title{Set the seed. The goal of this method is to centralise all calls to
the R method 'set.seed'.}
\usage{
setSeed(seed)
}
\arguments{
\item{seed}{seed value, not NULL}
}
\description{
Set the seed. The goal of this method is to centralise all calls to
the R method 'set.seed'.
}
