% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_setup.R
\name{setup_plan_default}
\alias{setup_plan_default}
\title{Setup default plan for the given simulation or hardware settings.
This plan will prioritise the distribution of workers in the following order:
1) Replicates (if 'replicate_parallel' is enabled)
2) Scenarios (if 'scenario_parallel' is enabled)
3) Dataset export / slices (if 'dataset_export' or 'slice_parallel' is enabled)}
\usage{
setup_plan_default(object)
}
\arguments{
\item{object}{simulation or hardware settings}
}
\value{
nothing
}
\description{
Setup default plan for the given simulation or hardware settings.
This plan will prioritise the distribution of workers in the following order:
1) Replicates (if 'replicate_parallel' is enabled)
2) Scenarios (if 'scenario_parallel' is enabled)
3) Dataset export / slices (if 'dataset_export' or 'slice_parallel' is enabled)
}
