% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{setupPlanDefault}
\alias{setupPlanDefault}
\title{Setup default plan for the given simulation or hardware settings.
This plan will prioritise the distribution of workers in the following order:
1) Replicates (if 'replicate_parallel' is enabled)
2) Scenarios (if 'scenario_parallel' is enabled)
3) Dataset export / slices (if 'dataset_export' or 'slice_parallel' is enabled)}
\usage{
setupPlanDefault(object)
}
\arguments{
\item{object}{simulation or hardware settings}
}
\value{
nothing
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`setupPlanDefault()` is deprecated in favor of `setup_plan_default()`.
}
\keyword{internal}
