% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_delegate}
\alias{simulate_delegate}
\title{Simulation delegate (several replicates).}
\usage{
simulate_delegate(
  model,
  dataset,
  dest,
  events,
  scenarios,
  tablefun,
  outvars,
  outfun,
  seed,
  replicates,
  dosing,
  settings
)
}
\arguments{
\item{model}{generic Campsis model}

\item{dataset}{Campsis dataset or 2-dimensional table}

\item{dest}{destination simulation engine, default is 'RxODE'}

\item{events}{interruption events}

\item{scenarios}{list of scenarios to be simulated}

\item{tablefun}{function or lambda formula to apply on exported 2-dimensional dataset}

\item{outvars}{variables to output in resulting dataframe}

\item{outfun}{an output function to apply on the simulation results. Type ?Outfun for more info.}

\item{seed}{seed value}

\item{replicates}{number of replicates, default is 1}

\item{dosing}{output dosing information, default is FALSE}

\item{settings}{advanced simulation settings}
}
\value{
a data frame with the results
}
\description{
Simulation delegate (several replicates).
}
\keyword{internal}
