% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_progress.R
\docType{class}
\name{simulation_progress-class}
\alias{simulation_progress-class}
\title{Simulation progress class.}
\description{
Simulation progress class.
}
\section{Slots}{

\describe{
\item{\code{replicates}}{total number of replicates to simulate}

\item{\code{scenarios}}{total number of scenarios to simulate}

\item{\code{iterations}}{total number of iterations to simulate}

\item{\code{slices}}{total number of slices to simulate}

\item{\code{replicate}}{current replicate number being simulated}

\item{\code{scenario}}{current scenario number being simulated}

\item{\code{iteration}}{current iteration number being simulated}

\item{\code{slice}}{current slice number being simulated}

\item{\code{progressor}}{progressr progressor}

\item{\code{hardware}}{hardware settings}
}}

