% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{sort}
\alias{sort}
\title{Sort the specified list.}
\usage{
sort(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{list object}

\item{decreasing}{increasing or decreasing order}

\item{...}{extra arguments}
}
\value{
same list but ordered
}
\description{
Sort the specified list.
}
