% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaguetti_plot.R
\name{spaguetti_plot}
\alias{spaguetti_plot}
\title{Spaghetti plot (S3 generic).}
\usage{
spaguetti_plot(x, ...)
}
\arguments{
\item{x}{a CAMPSIS output object}

\item{...}{additional arguments passed to the method}
}
\value{
a ggplot object
}
\description{
Spaghetti plot (S3 generic).
}
