% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{spaguettiPlot}
\alias{spaguettiPlot}
\title{Spaguetti plot.}
\usage{
spaguettiPlot(x, output, scenarios = NULL)
}
\arguments{
\item{x}{data frame}

\item{output}{variable to show}

\item{scenarios}{scenarios}
}
\description{
Spaguetti plot.
}
