% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{splitDataset}
\alias{splitDataset}
\title{Split dataset according to config.}
\usage{
splitDataset(dataset, config)
}
\arguments{
\item{dataset}{Campsis dataset to export}

\item{config}{current iteration in future_map_dfr}
}
\value{
a subset of the given dataset
}
\description{
Split dataset according to config.
}
\keyword{internal}
