% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utilities.R
\name{standardiseTime}
\alias{standardiseTime}
\title{Standardise time to hours.}
\usage{
standardiseTime(x, unit)
}
\arguments{
\item{x}{numeric time vector}

\item{unit}{unit of x, single character value}
}
\value{
numeric vector with the times converted to hours
}
\description{
Standardise time to hours.
}
