% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun.R
\docType{class}
\name{stats_outfun-class}
\alias{stats_outfun-class}
\title{Statistics output function class.}
\description{
Statistics output function class.
}
\section{Slots}{

\describe{
\item{\code{variable}}{variable(s) used to compute the statistics, character vector}

\item{\code{strata}}{named vector with the strata to use}

\item{\code{stats}}{character vector of statistics to compute}
}}

