% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun.R
\name{StatsOutfun}
\alias{StatsOutfun}
\title{Create a statistics output function}
\usage{
StatsOutfun(
  variable,
  strata = get_default_strata(),
  stats = c("p5", "median", "p95"),
  name = sprintf("stats_\%s", paste0(variable, collapse = "_"))
)
}
\arguments{
\item{variable}{variable(s) used to compute the statistics, character vector}

\item{strata}{named vector with the strata to use, default is c(SCENARIO="all", ARM="all")}

\item{stats}{character vector of statistics to compute. Supported: "median", "mean", or percentiles like "p5", "p95", "p2.5", "p97.5". Default is c("p5", "median", "p95").}

\item{name}{name of the output function. Default is 'stats_<variable>'.}
}
\value{
a stats_outfun object
}
\description{
Create a statistics output function
}
