% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_vector.R
\name{TimeSequence}
\alias{TimeSequence}
\title{Instantiate a new time sequence.}
\usage{
TimeSequence(start, end, by)
}
\arguments{
\item{start}{starting value}

\item{end}{maximal value}

\item{by}{increment of the sequence}
}
\value{
a sequence
}
\description{
Instantiate a new time sequence.
}
