% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_vector.R
\name{TimeVector}
\alias{TimeVector}
\title{Instantiate a new time vector}
\usage{
TimeVector(x)
}
\arguments{
\item{x}{time vector, numeric}
}
\value{
a time vector
}
\description{
Instantiate a new time vector
}
