% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_summary.R
\name{toDatasetSummary}
\alias{toDatasetSummary}
\title{Convert dataset to dataset summary (internal method).}
\usage{
toDatasetSummary(dataset)
}
\value{
dataset summary
}
\description{
Convert dataset to dataset summary (internal method).
}
\keyword{internal}
