% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{toExplicitDistribution}
\alias{toExplicitDistribution}
\title{Convert user-given distribution to an explicit pmxsim distribution.
Passed distribution can be:
- a NULL value. In that case, it will be converted into is a 'UndefinedDistribution'.
- a single numeric value. In that case, it will be converted into is a 'ConstantDistribution'.
- a numeric vector. In that case, it will be converted into is a 'FixedDistribution'.
- all available types of distribution. In this case, no conversion is applied.}
\usage{
toExplicitDistribution(distribution)
}
\arguments{
\item{distribution}{user-given distribution}
}
\value{
a distribution object
}
\description{
Convert user-given distribution to an explicit pmxsim distribution.
Passed distribution can be:
- a NULL value. In that case, it will be converted into is a 'UndefinedDistribution'.
- a single numeric value. In that case, it will be converted into is a 'ConstantDistribution'.
- a numeric vector. In that case, it will be converted into is a 'FixedDistribution'.
- all available types of distribution. In this case, no conversion is applied.
}
\keyword{internal}
