% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_bioavailability.R
\name{TreatmentBioavailability}
\alias{TreatmentBioavailability}
\title{Create a bioavailability for the specified compartment.}
\usage{
TreatmentBioavailability(compartment, distribution)
}
\arguments{
\item{compartment}{compartment number}

\item{distribution}{distribution}
}
\value{
bioavailability
}
\description{
Create a bioavailability for the specified compartment.
}
