% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_lag_time.R
\name{TreatmentLagTime}
\alias{TreatmentLagTime}
\title{Create a lag time for the specified compartment.}
\usage{
TreatmentLagTime(compartment, distribution)
}
\arguments{
\item{compartment}{compartment number}

\item{distribution}{distribution}
}
\value{
lag time
}
\description{
Create a lag time for the specified compartment.
}
