% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{UniformDistribution}
\alias{UniformDistribution}
\title{Create an uniform distribution.}
\usage{
UniformDistribution(min, max)
}
\arguments{
\item{min}{min value}

\item{max}{max value}
}
\value{
an uniform distribution
}
\description{
Create an uniform distribution.
}
