% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{uniteColumns}
\alias{uniteColumns}
\title{Unite the given column names.}
\usage{
uniteColumns(x, columns, colname, factor = TRUE)
}
\arguments{
\item{x}{data frame, CAMPSIS output}

\item{columns}{columns to unify}

\item{colname}{destination column name}

\item{factor}{factor the destination column}
}
\value{
a data frame
}
\description{
Unite the given column names.
}
\keyword{internal}
