% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment_entry.R, R/treatment.R,
%   R/arm.R, R/arms.R, R/dataset.R
\name{unwrapTreatment}
\alias{unwrapTreatment}
\alias{unwrapTreatment,bolus-method}
\alias{unwrapTreatment,infusion-method}
\alias{unwrapTreatment,bolus_wrapper-method}
\alias{unwrapTreatment,infusion_wrapper-method}
\alias{unwrapTreatment,treatment-method}
\alias{unwrapTreatment,arm-method}
\alias{unwrapTreatment,arms-method}
\alias{unwrapTreatment,dataset-method}
\title{Unwrap treatment.}
\usage{
unwrapTreatment(object)

\S4method{unwrapTreatment}{bolus}(object)

\S4method{unwrapTreatment}{infusion}(object)

\S4method{unwrapTreatment}{bolus_wrapper}(object)

\S4method{unwrapTreatment}{infusion_wrapper}(object)

\S4method{unwrapTreatment}{treatment}(object)

\S4method{unwrapTreatment}{arm}(object)

\S4method{unwrapTreatment}{arms}(object)

\S4method{unwrapTreatment}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
updated object
}
\description{
Unwrap treatment.
}
