% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment_entry.R, R/treatment.R,
%   R/arm.R, R/arms.R, R/dataset.R
\name{update_addl}
\alias{update_addl}
\alias{update_addl,bolus_wrapper,integer,character-method}
\alias{update_addl,infusion_wrapper,integer,character-method}
\alias{update_addl,bolus,integer,character-method}
\alias{update_addl,infusion,integer,character-method}
\alias{update_addl,treatment,integer,character-method}
\alias{update_addl,arm,integer,character-method}
\alias{update_addl,arms,integer,character-method}
\alias{update_addl,dataset,integer,character-method}
\title{Update the number of additional doses (ADDL).}
\usage{
update_addl(object, addl, ref = NULL)

\S4method{update_addl}{bolus_wrapper,integer,character}(object, addl, ref = NULL)

\S4method{update_addl}{infusion_wrapper,integer,character}(object, addl, ref = NULL)

\S4method{update_addl}{bolus,integer,character}(object, addl, ref = NULL)

\S4method{update_addl}{infusion,integer,character}(object, addl, ref = NULL)

\S4method{update_addl}{treatment,integer,character}(object, addl, ref = NULL)

\S4method{update_addl}{arm,integer,character}(object, addl, ref = NULL)

\S4method{update_addl}{arms,integer,character}(object, addl, ref = NULL)

\S4method{update_addl}{dataset,integer,character}(object, addl, ref = NULL)
}
\arguments{
\item{object}{generic object}

\item{addl}{new number of additional doses}

\item{ref}{reference treatment name}
}
\value{
updated object
}
\description{
Update the number of additional doses (ADDL).
}
