% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment_entry.R, R/treatment.R,
%   R/arm.R, R/arms.R, R/dataset.R
\name{update_amount}
\alias{update_amount}
\alias{update_amount,bolus,numeric,character-method}
\alias{update_amount,infusion,numeric,character-method}
\alias{update_amount,bolus_wrapper,numeric,character-method}
\alias{update_amount,infusion_wrapper,numeric,character-method}
\alias{update_amount,treatment,numeric,character-method}
\alias{update_amount,arm,numeric,character-method}
\alias{update_amount,arms,numeric,character-method}
\alias{update_amount,dataset,numeric,character-method}
\title{Update amount.}
\usage{
update_amount(object, amount, ref)

\S4method{update_amount}{bolus,numeric,character}(object, amount, ref)

\S4method{update_amount}{infusion,numeric,character}(object, amount, ref)

\S4method{update_amount}{bolus_wrapper,numeric,character}(object, amount, ref)

\S4method{update_amount}{infusion_wrapper,numeric,character}(object, amount, ref)

\S4method{update_amount}{treatment,numeric,character}(object, amount, ref)

\S4method{update_amount}{arm,numeric,character}(object, amount, ref)

\S4method{update_amount}{arms,numeric,character}(object, amount, ref)

\S4method{update_amount}{dataset,numeric,character}(object, amount, ref)
}
\arguments{
\item{object}{generic object}

\item{amount}{new amount}

\item{ref}{reference treatment name}
}
\value{
updated object
}
\description{
Update amount.
}
