% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment_entry.R, R/treatment.R,
%   R/arm.R, R/arms.R, R/dataset.R
\name{update_repeat}
\alias{update_repeat}
\alias{update_repeat,bolus_wrapper,repeated_schedule,character-method}
\alias{update_repeat,infusion_wrapper,repeated_schedule,character-method}
\alias{update_repeat,bolus,repeated_schedule,character-method}
\alias{update_repeat,infusion,repeated_schedule,character-method}
\alias{update_repeat,treatment,repeated_schedule,character-method}
\alias{update_repeat,arm,repeated_schedule,character-method}
\alias{update_repeat,arms,repeated_schedule,character-method}
\alias{update_repeat,dataset,repeated_schedule,character-method}
\title{Update the repeat field (argument 'rep' in Bolus and Infusion constructors).}
\usage{
update_repeat(object, rep, ref = NULL)

\S4method{update_repeat}{bolus_wrapper,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{update_repeat}{infusion_wrapper,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{update_repeat}{bolus,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{update_repeat}{infusion,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{update_repeat}{treatment,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{update_repeat}{arm,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{update_repeat}{arms,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{update_repeat}{dataset,repeated_schedule,character}(object, rep, ref = NULL)
}
\arguments{
\item{object}{generic object}

\item{rep}{repeated dosing schedule (definition) object}

\item{ref}{reference treatment name}
}
\value{
updated object
}
\description{
Update the repeat field (argument 'rep' in Bolus and Infusion constructors).
}
