% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment_entry.R, R/treatment.R,
%   R/arm.R, R/arms.R, R/dataset.R
\name{updateADDL}
\alias{updateADDL}
\alias{updateADDL,bolus_wrapper,integer,character-method}
\alias{updateADDL,infusion_wrapper,integer,character-method}
\alias{updateADDL,bolus,integer,character-method}
\alias{updateADDL,infusion,integer,character-method}
\alias{updateADDL,treatment,integer,character-method}
\alias{updateADDL,arm,integer,character-method}
\alias{updateADDL,arms,integer,character-method}
\alias{updateADDL,dataset,integer,character-method}
\title{Update the number of additional doses (ADDL).}
\usage{
updateADDL(object, addl, ref = NULL)

\S4method{updateADDL}{bolus_wrapper,integer,character}(object, addl, ref = NULL)

\S4method{updateADDL}{infusion_wrapper,integer,character}(object, addl, ref = NULL)

\S4method{updateADDL}{bolus,integer,character}(object, addl, ref = NULL)

\S4method{updateADDL}{infusion,integer,character}(object, addl, ref = NULL)

\S4method{updateADDL}{treatment,integer,character}(object, addl, ref = NULL)

\S4method{updateADDL}{arm,integer,character}(object, addl, ref = NULL)

\S4method{updateADDL}{arms,integer,character}(object, addl, ref = NULL)

\S4method{updateADDL}{dataset,integer,character}(object, addl, ref = NULL)
}
\arguments{
\item{object}{generic object}

\item{addl}{new number of additional doses}

\item{ref}{reference treatment name}
}
\value{
updated object
}
\description{
Update the number of additional doses (ADDL).
}
