% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment_entry.R, R/treatment.R,
%   R/arm.R, R/arms.R, R/dataset.R
\name{updateAmount}
\alias{updateAmount}
\alias{updateAmount,bolus,numeric,character-method}
\alias{updateAmount,infusion,numeric,character-method}
\alias{updateAmount,bolus_wrapper,numeric,character-method}
\alias{updateAmount,infusion_wrapper,numeric,character-method}
\alias{updateAmount,treatment,numeric,character-method}
\alias{updateAmount,arm,numeric,character-method}
\alias{updateAmount,arms,numeric,character-method}
\alias{updateAmount,dataset,numeric,character-method}
\title{Update amount.}
\usage{
updateAmount(object, amount, ref)

\S4method{updateAmount}{bolus,numeric,character}(object, amount, ref)

\S4method{updateAmount}{infusion,numeric,character}(object, amount, ref)

\S4method{updateAmount}{bolus_wrapper,numeric,character}(object, amount, ref)

\S4method{updateAmount}{infusion_wrapper,numeric,character}(object, amount, ref)

\S4method{updateAmount}{treatment,numeric,character}(object, amount, ref)

\S4method{updateAmount}{arm,numeric,character}(object, amount, ref)

\S4method{updateAmount}{arms,numeric,character}(object, amount, ref)

\S4method{updateAmount}{dataset,numeric,character}(object, amount, ref)
}
\arguments{
\item{object}{generic object}

\item{amount}{new amount}

\item{ref}{reference treatment name}
}
\value{
updated object
}
\description{
Update amount.
}
