% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment_entry.R, R/treatment.R,
%   R/arm.R, R/arms.R, R/dataset.R
\name{updateDoseInterval}
\alias{updateDoseInterval}
\alias{updateDoseInterval,bolus_wrapper,numeric,integer,character-method}
\alias{updateDoseInterval,infusion_wrapper,numeric,integer,character-method}
\alias{updateDoseInterval,bolus,numeric,integer,character-method}
\alias{updateDoseInterval,infusion,numeric,integer,character-method}
\alias{updateDoseInterval,treatment,numeric,integer,character-method}
\alias{updateDoseInterval,arm,numeric,integer,character-method}
\alias{updateDoseInterval,arms,numeric,integer,character-method}
\alias{updateDoseInterval,dataset,numeric,integer,character-method}
\title{Update dose interval.}
\usage{
updateDoseInterval(object, ii = NULL, addl = NULL, ref = NULL)

\S4method{updateDoseInterval}{bolus_wrapper,numeric,integer,character}(object, ii = NULL, addl = NULL, ref = NULL)

\S4method{updateDoseInterval}{infusion_wrapper,numeric,integer,character}(object, ii = NULL, addl = NULL, ref = NULL)

\S4method{updateDoseInterval}{bolus,numeric,integer,character}(object, ii = NULL, addl = NULL, ref = NULL)

\S4method{updateDoseInterval}{infusion,numeric,integer,character}(object, ii = NULL, addl = NULL, ref = NULL)

\S4method{updateDoseInterval}{treatment,numeric,integer,character}(object, ii = NULL, addl = NULL, ref = NULL)

\S4method{updateDoseInterval}{arm,numeric,integer,character}(object, ii = NULL, addl = NULL, ref = NULL)

\S4method{updateDoseInterval}{arms,numeric,integer,character}(object, ii = NULL, addl = NULL, ref = NULL)

\S4method{updateDoseInterval}{dataset,numeric,integer,character}(object, ii = NULL, addl = NULL, ref = NULL)
}
\arguments{
\item{object}{generic object}

\item{ii}{new inter-dose interval}

\item{addl}{new number of additional doses}

\item{ref}{reference treatment name}
}
\value{
updated object
}
\description{
Update dose interval.
}
