% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment_entry.R, R/treatment.R,
%   R/arm.R, R/arms.R, R/dataset.R
\name{updateII}
\alias{updateII}
\alias{updateII,bolus_wrapper,numeric,character-method}
\alias{updateII,infusion_wrapper,numeric,character-method}
\alias{updateII,bolus,numeric,character-method}
\alias{updateII,infusion,numeric,character-method}
\alias{updateII,treatment,numeric,character-method}
\alias{updateII,arm,numeric,character-method}
\alias{updateII,arms,numeric,character-method}
\alias{updateII,dataset,numeric,character-method}
\title{Update the inter-dose interval (II).}
\usage{
updateII(object, ii, ref = NULL)

\S4method{updateII}{bolus_wrapper,numeric,character}(object, ii, ref = NULL)

\S4method{updateII}{infusion_wrapper,numeric,character}(object, ii, ref = NULL)

\S4method{updateII}{bolus,numeric,character}(object, ii, ref = NULL)

\S4method{updateII}{infusion,numeric,character}(object, ii, ref = NULL)

\S4method{updateII}{treatment,numeric,character}(object, ii, ref = NULL)

\S4method{updateII}{arm,numeric,character}(object, ii, ref = NULL)

\S4method{updateII}{arms,numeric,character}(object, ii, ref = NULL)

\S4method{updateII}{dataset,numeric,character}(object, ii, ref = NULL)
}
\arguments{
\item{object}{generic object}

\item{ii}{new inter-dose interval}

\item{ref}{reference treatment name}
}
\value{
updated object
}
\description{
Update the inter-dose interval (II).
}
