% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment_entry.R, R/treatment.R,
%   R/arm.R, R/arms.R, R/dataset.R
\name{updateRepeat}
\alias{updateRepeat}
\alias{updateRepeat,bolus_wrapper,repeated_schedule,character-method}
\alias{updateRepeat,infusion_wrapper,repeated_schedule,character-method}
\alias{updateRepeat,bolus,repeated_schedule,character-method}
\alias{updateRepeat,infusion,repeated_schedule,character-method}
\alias{updateRepeat,treatment,repeated_schedule,character-method}
\alias{updateRepeat,arm,repeated_schedule,character-method}
\alias{updateRepeat,arms,repeated_schedule,character-method}
\alias{updateRepeat,dataset,repeated_schedule,character-method}
\title{Update the repeat field (argument 'rep' in Bolus and Infusion constructors).}
\usage{
updateRepeat(object, rep, ref = NULL)

\S4method{updateRepeat}{bolus_wrapper,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{updateRepeat}{infusion_wrapper,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{updateRepeat}{bolus,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{updateRepeat}{infusion,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{updateRepeat}{treatment,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{updateRepeat}{arm,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{updateRepeat}{arms,repeated_schedule,character}(object, rep, ref = NULL)

\S4method{updateRepeat}{dataset,repeated_schedule,character}(object, rep, ref = NULL)
}
\arguments{
\item{object}{generic object}

\item{rep}{repeated dosing schedule (definition) object}

\item{ref}{reference treatment name}
}
\value{
updated object
}
\description{
Update the repeat field (argument 'rep' in Bolus and Infusion constructors).
}
