% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utilities.R
\name{years}
\alias{years}
\title{Convert pharma years (1 year = 12*4 weeks) to hours.}
\usage{
years(x)
}
\arguments{
\item{x}{numeric vector in years}
}
\value{
numeric vector in hours
}
\description{
Convert pharma years (1 year = 12*4 weeks) to hours.
}
