% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{apply_outfun}
\alias{apply_outfun}
\title{Apply output function(s) on the given Campsis results.}
\usage{
apply_outfun(x, outfun, level, ...)
}
\arguments{
\item{x}{Campsis simulation results}

\item{outfun}{output function(s), an `outfun` or `outfuns` object}

\item{level}{level at which the output function is applied, e.g. 'replicate'}

\item{...}{extra arguments transmitted automatically by Campsis (e.g. `replicate`, `scenario`)}
}
\value{
the updated Campsis results
}
\description{
Apply output function(s) on the given Campsis results.
}
