% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/scenario_action.R
\name{applyAction}
\alias{applyAction}
\alias{applyAction,campsis_model,replace_action-method}
\alias{applyAction,dataset,replace_action-method}
\title{Apply some action on the given object.}
\usage{
applyAction(object, action)

\S4method{applyAction}{campsis_model,replace_action}(object, action)

\S4method{applyAction}{dataset,replace_action}(object, action)
}
\arguments{
\item{object}{any object}

\item{action}{action to apply}
}
\value{
updated object
}
\description{
Apply some action on the given object.
}
