% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{applyCompartmentCharacteristics}
\alias{applyCompartmentCharacteristics}
\title{Apply compartment characteristics from model.}
\usage{
applyCompartmentCharacteristics(table, characteristics)
}
\arguments{
\item{table}{current dataset}

\item{characteristics}{compartment characteristics from model}
}
\value{
updated dataset
}
\description{
Apply compartment characteristics from model.
}
