% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.R
\name{Arm}
\alias{Arm}
\title{Create a treatment arm.}
\usage{
Arm(id = as.integer(NA), subjects = 1, label = as.character(NA))
}
\arguments{
\item{id}{unique identifier for this arm (available trough dataset), integer. If NA (default), this identifier is auto-incremented.}

\item{subjects}{number of subjects in arm, integer}

\item{label}{arm label, single character string. If set, this label will be output in the ARM column of Campsis instead of the identifier.}
}
\value{
an arm
}
\description{
Create a treatment arm.
}
