% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.R
\name{Arm}
\alias{Arm}
\title{Create a treatment arm.}
\usage{
Arm(id = 1, subjects = 1)
}
\arguments{
\item{id}{arm unique ID, integer}

\item{subjects}{number of subjects in arm, integer}
}
\value{
an arm
}
\description{
Create a treatment arm.
}
