% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_entry.R
\name{Bolus}
\alias{Bolus}
\title{Create a bolus.}
\usage{
Bolus(time, amount, compartment = NA, fraction = NULL, lag = NULL)
}
\arguments{
\item{time}{treatment time, numeric}

\item{amount}{amount to give as bolus, numeric}

\item{compartment}{compartment index, integer}

\item{fraction}{fraction of dose, distribution}

\item{lag}{dose lag time, distribution}
}
\value{
an observation
}
\description{
Create a bolus.
}
