% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{Bootstrap}
\alias{Bootstrap}
\title{Create a bootstrap object.}
\usage{
Bootstrap(data, replacement = FALSE, random = FALSE, output_id = FALSE)
}
\arguments{
\item{replacement}{values can be reused or not, logical}

\item{random}{values are drawn randomly, logical}

\item{output_id}{tell CAMPSIS if 'BS_ID' must be output or not, logical}
}
\value{
a covariate
}
\description{
Create a bootstrap object.
}
\section{Slots}{

\describe{
\item{\code{data}}{data frame to be bootstrapped. Column 'BS_ID' is mandatory and
corresponds to the original ID from the bootstrap. It must be numeric and unique.
Other columns are covariates to be bootstrapped (row by row).}
}}

