% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{BootstrapDistribution}
\alias{BootstrapDistribution}
\title{Create a bootstrap distribution. During function sampling, PMXsim will generate
values depending on the given data and arguments.}
\usage{
BootstrapDistribution(data, replacement = FALSE, random = FALSE)
}
\arguments{
\item{data}{values to draw, numeric vector}

\item{replacement}{values can be reused or not, logical}

\item{random}{values are drawn randomly, logical}
}
\value{
a covariate
}
\description{
Create a bootstrap distribution. During function sampling, PMXsim will generate
values depending on the given data and arguments.
}
