% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_progress.R
\name{campsis_handler}
\alias{campsis_handler}
\title{Suggested Campsis handler for showing the progress bar.}
\usage{
campsis_handler()
}
\value{
a progressr handler list
}
\description{
Suggested Campsis handler for showing the progress bar.
}
