% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_progress.R
\name{compute_incremental_progress}
\alias{compute_incremental_progress}
\title{Compute incremental progress.}
\usage{
compute_incremental_progress(object, tick_slice)
}
\arguments{
\item{object}{simulation progress object}

\item{tick_slice}{tick progress on slices}
}
\value{
incremental progress, in percentage
}
\description{
Compute incremental progress.
}
\keyword{internal}
