% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun_core.R
\name{compute_pi}
\alias{compute_pi}
\title{Compute the prediction interval summary over time.}
\usage{
compute_pi(x, variable, strata = get_default_strata(), level = 0.9)
}
\arguments{
\item{x}{data frame}

\item{variable}{variable(s) used to compute the prediction interval, character vector.}

\item{strata}{named vector with the strata to use, default is c(SCENARIO="all", ARM="all").}

\item{level}{PI level, default is 0.9 (90\% PI)}
}
\value{
a summary table
}
\description{
Compute the prediction interval summary over time.
}
