% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_progress.R
\name{computeIncrementalProgress}
\alias{computeIncrementalProgress}
\title{Compute incremental progress.}
\usage{
computeIncrementalProgress(object, tick_slice)
}
\arguments{
\item{object}{simulation progress object}

\item{tick_slice}{tick progress on slices}
}
\value{
incremental progress, in percentage
}
\description{
Compute incremental progress.
}
\keyword{internal}
