% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{convert}
\alias{convert}
\title{Convert dataset entry.}
\usage{
convert(object)
}
\arguments{
\item{object}{generic object}

\item{entry}{entry to add}
}
\value{
generic object
}
\description{
Convert dataset entry.
}
