% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{convertTime}
\alias{convertTime}
\title{Convert numeric time vector based on the provided units.}
\usage{
convertTime(x, from, to)
}
\arguments{
\item{x}{numeric time vector}

\item{from}{unit of x, single character value}

\item{to}{destination unit, single character value}
}
\value{
numeric vector with the converted times
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`convertTime()` is deprecated in favor of `convert_time()`.
}
\keyword{internal}
