% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{Covariate}
\alias{Covariate}
\title{Create a fixed covariate.}
\usage{
Covariate(name, distribution)
}
\arguments{
\item{name}{covariate name, character}

\item{distribution}{covariate distribution}
}
\value{
a fixed covariate
}
\description{
Create a fixed covariate.
}
