% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_logic.R
\name{cutTableForEvent}
\alias{cutTableForEvent}
\title{Cut table according to given iteration.}
\usage{
cutTableForEvent(table, iteration)
}
\arguments{
\item{table}{whole table, data frame}

\item{iteration}{current iteration being processed}
}
\description{
Cut table according to given iteration.
}
\keyword{internal}
