% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{cutTableForEvent}
\alias{cutTableForEvent}
\title{Cut table for event starting at 'start' and ending at 'end'.}
\usage{
cutTableForEvent(table, current_event)
}
\arguments{
\item{table}{whole table, data frame}

\item{current_event}{current event being processed}
}
\description{
Cut table for event starting at 'start' and ending at 'end'.
}
\keyword{internal}
