% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset}
\alias{Dataset}
\title{Create a dataset.}
\usage{
Dataset(subjects = NULL, label = as.character(NA), json = NULL)
}
\arguments{
\item{subjects}{number of subjects in the default arm}

\item{label}{label of the default arm, NA by default}

\item{json}{path to JSON dataset file or JSON content in string form}
}
\value{
a dataset
}
\description{
Create a dataset.
}
