% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset}
\alias{Dataset}
\title{Create a dataset.}
\usage{
Dataset(subjects = NULL)
}
\arguments{
\item{subjects}{number of subjects in the default arm}
}
\value{
a dataset
}
\description{
Create a dataset.
}
