% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_config.R
\docType{class}
\name{dataset_config-class}
\alias{dataset_config-class}
\title{Dataset configuration class.}
\description{
Dataset configuration class.
}
\section{Slots}{

\describe{
\item{\code{def_depot_cmt}}{default depot compartment, integer}

\item{\code{def_obs_cmt}}{default observation compartment, integer}
}}

