% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_config.R
\name{DatasetConfig}
\alias{DatasetConfig}
\title{Create dataset configuration.}
\usage{
DatasetConfig(defDepotCmt = 1, defObsCmt = 1)
}
\arguments{
\item{defDepotCmt}{default depot compartment, integer}

\item{defObsCmt}{default observation compartment, integer}
}
\value{
configuration
}
\description{
Create dataset configuration.
}
