% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_settings.R
\docType{class}
\name{default_settings-class}
\alias{default_settings-class}
\title{Default settings class.}
\description{
Default settings class.
}
\section{Slots}{

\describe{
\item{\code{engine}}{simulation engine, character}

\item{\code{seed}}{random seed number, integer}

\item{\code{outvars}}{output variables, character vector}

\item{\code{disabled_variabilities}}{variabilities to disable in the simulation, character vector}

\item{\code{dosing}}{output dosing information, logical}
}}

