% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_settings.R
\name{DefaultSettings}
\alias{DefaultSettings}
\title{Create default settings.}
\usage{
DefaultSettings(
  engine = "rxode2",
  seed = NULL,
  outvars = character(),
  outfuns = Outfuns(),
  disabled_variabilities = character(),
  dosing = FALSE
)
}
\arguments{
\item{engine}{simulation engine, character}

\item{seed}{random seed number, integer (or NULL for auto-generated seed)}

\item{outvars}{output variables, character vector}

\item{outfuns}{output functions, outfuns object}

\item{disabled_variabilities}{variabilities to disable in the simulation, character vector}

\item{dosing}{output dosing information, logical}
}
\value{
default settings
}
\description{
Create default settings.
}
