% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_settings.R
\name{DefaultSettings}
\alias{DefaultSettings}
\title{Create default settings.}
\usage{
DefaultSettings(
  engine = NULL,
  seed = NULL,
  replicates = 1L,
  outvars = character(),
  outfuns = Outfuns(),
  disabled_variabilities = character(),
  dosing = FALSE
)
}
\arguments{
\item{engine}{Simulation engine: \code{'rxode2'} or \code{'mrgsolve'}.
Default is \code{NULL} (unspecified). If \code{NULL}, \code{'mrgsolve'} is used first
(if installed), followed by \code{'rxode2'} (if installed).}

\item{seed}{random seed number, integer (or NULL for auto-generated seed)}

\item{replicates}{number of replicates, integer. Default is 1.}

\item{outvars}{output variables, character vector}

\item{outfuns}{output functions, outfuns object}

\item{disabled_variabilities}{variabilities to disable in the simulation, character vector}

\item{dosing}{output dosing information, logical}
}
\value{
default settings
}
\description{
Create default settings.
}
