% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{dosing_only}
\alias{dosing_only}
\title{Filter Campsis output on dosing rows.}
\usage{
dosing_only(x)
}
\arguments{
\item{x}{data frame, Campsis output}
}
\value{
a data frame with the dosing rows
}
\description{
Filter Campsis output on dosing rows.
}
