% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{EtaDistribution}
\alias{EtaDistribution}
\title{Create an ETA distribution. The resulting distribution is a
normal distribution, with mean=0 and sd=sqrt(OMEGA).}
\usage{
EtaDistribution(model, omega)
}
\arguments{
\item{model}{model}

\item{omega}{corresponding THETA name, character}
}
\value{
an ETA distribution
}
\description{
Create an ETA distribution. The resulting distribution is a
normal distribution, with mean=0 and sd=sqrt(OMEGA).
}
