% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{generate_iiv_}
\alias{generate_iiv_}
\title{Generate IIV matrix for the given OMEGA matrix.}
\usage{
generate_iiv_(omega, n)
}
\arguments{
\item{omega}{omega matrix}

\item{n}{number of subjects}
}
\value{
IIV data frame
}
\description{
Generate IIV matrix for the given OMEGA matrix.
}
