% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{generateIIV}
\alias{generateIIV}
\title{Generate IIV in dataset.}
\usage{
generateIIV(dataset, model)
}
\arguments{
\item{dataset}{dataset}

\item{model}{model}
}
\value{
dataset
}
\description{
Generate IIV in dataset.
}
