% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/covariates.R, R/arm.R, R/arms.R,
%   R/dataset.R
\name{get_covariates}
\alias{get_covariates}
\alias{get_covariates,covariates-method}
\alias{get_covariates,arm-method}
\alias{get_covariates,arms-method}
\alias{get_covariates,dataset-method}
\title{Get all covariates (fixed / time-varying / event covariates).}
\usage{
get_covariates(object)

\S4method{get_covariates}{covariates}(object)

\S4method{get_covariates}{arm}(object)

\S4method{get_covariates}{arms}(object)

\S4method{get_covariates}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
all covariates from object
}
\description{
Get all covariates (fixed / time-varying / event covariates).
}
