% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/covariates.R, R/arm.R, R/arms.R,
%   R/dataset.R
\name{get_event_covariates}
\alias{get_event_covariates}
\alias{get_event_covariates,covariates-method}
\alias{get_event_covariates,arm-method}
\alias{get_event_covariates,arms-method}
\alias{get_event_covariates,dataset-method}
\title{Get all event-related covariates.}
\usage{
get_event_covariates(object)

\S4method{get_event_covariates}{covariates}(object)

\S4method{get_event_covariates}{arm}(object)

\S4method{get_event_covariates}{arms}(object)

\S4method{get_event_covariates}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
all event-related covariates from object
}
\description{
Get all event-related covariates.
}
