% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/arm.R, R/arms.R, R/dataset.R
\name{get_iovs}
\alias{get_iovs}
\alias{get_iovs,arm-method}
\alias{get_iovs,arms-method}
\alias{get_iovs,dataset-method}
\title{Get all IOV objects.}
\usage{
get_iovs(object)

\S4method{get_iovs}{arm}(object)

\S4method{get_iovs}{arms}(object)

\S4method{get_iovs}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
all IOV's from object
}
\description{
Get all IOV objects.
}
