% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{get_random_seed_value}
\alias{get_random_seed_value}
\title{Get random seed value.}
\usage{
get_random_seed_value()
}
\value{
random seed value generated based on time
}
\description{
Get random seed value.
}
\keyword{internal}
