% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{get_seed_for_dataset_export}
\alias{get_seed_for_dataset_export}
\title{Get seed for dataset export.}
\usage{
get_seed_for_dataset_export(seed, progress)
}
\arguments{
\item{seed}{original seed}

\item{progress}{simulation progress}
}
\value{
the seed value used to export the dataset
}
\description{
Get seed for dataset export.
}
\keyword{internal}
