% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{get_seed_for_iteration}
\alias{get_seed_for_iteration}
\title{Get seed for iteration.}
\usage{
get_seed_for_iteration(seed, progress)
}
\arguments{
\item{seed}{original seed}

\item{progress}{simulation progress}
}
\value{
the seed value to be used for the given replicate number and iteration
}
\description{
Get seed for iteration.
}
\keyword{internal}
