% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_splitting_configuration}
\alias{get_splitting_configuration}
\title{Get splitting configuration for parallel export.}
\usage{
get_splitting_configuration(dataset, hardware)
}
\arguments{
\item{dataset}{Campsis dataset to export}

\item{hardware}{hardware configuration}
}
\value{
splitting configuration list (if 'dataset_parallel' is enabled) or
 NA (if 'dataset_parallel' disabled or if the length of the dataset is less than the dataset export slice size)
}
\description{
Get splitting configuration for parallel export.
}
