% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{getByName}
\alias{getByName}
\title{Get an element from a list by name.}
\usage{
getByName(object, name)
}
\arguments{
\item{object}{list object}

\item{name}{element name to search for}
}
\value{
index of this element
}
\description{
Get an element from a list by name.
}
