% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/arm.R, R/arms.R, R/dataset.R
\name{getCovariateNames}
\alias{getCovariateNames}
\alias{getCovariateNames,arm-method}
\alias{getCovariateNames,arms-method}
\alias{getCovariateNames,dataset-method}
\title{Get all covariate names (fixed covariates + time-varying covariates).}
\usage{
getCovariateNames(object)

\S4method{getCovariateNames}{arm}(object)

\S4method{getCovariateNames}{arms}(object)

\S4method{getCovariateNames}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
character vector
}
\description{
Get all covariate names (fixed covariates + time-varying covariates).
}
