% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_logic.R
\name{getEventIterations}
\alias{getEventIterations}
\title{Get list of event iterations.}
\usage{
getEventIterations(events, dataset)
}
\arguments{
\item{events}{events}

\item{dataset}{dataset, Campsis dataset or data frame}
}
\value{
a list of event iterations. Note: start and end slots have the same unit as the 'exported' dataset time.
}
\description{
Get list of event iterations.
}
\keyword{internal}
