% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_logic.R
\name{getEventIterations}
\alias{getEventIterations}
\title{Get list of event iterations.}
\usage{
getEventIterations(events, maxTime)
}
\arguments{
\item{events}{events}

\item{maxTime}{simulation max time}
}
\value{
a list of event iterations
}
\description{
Get list of event iterations.
}
\keyword{internal}
