% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_list.R
\name{getNames}
\alias{getNames}
\title{Get names of element in list.}
\usage{
getNames(object)
}
\arguments{
\item{object}{list object}
}
\value{
character vector
}
\description{
Get names of element in list.
}
