% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/arm.R, R/arms.R, R/dataset.R
\name{getOccasionNames}
\alias{getOccasionNames}
\alias{getOccasionNames,arm-method}
\alias{getOccasionNames,arms-method}
\alias{getOccasionNames,dataset-method}
\title{Get all occasion names.}
\usage{
getOccasionNames(object)

\S4method{getOccasionNames}{arm}(object)

\S4method{getOccasionNames}{arms}(object)

\S4method{getOccasionNames}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
character vector
}
\description{
Get all occasion names.
}
