% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{getSeed}
\alias{getSeed}
\title{Get seed value.}
\usage{
getSeed(seed = NULL)
}
\arguments{
\item{seed}{user-input seed. If NULL or NA, seed number will be random.}
}
\value{
a seed value, integer
}
\description{
Get seed value.
}
\keyword{internal}
