% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{getSeedForDatasetExport}
\alias{getSeedForDatasetExport}
\title{Get seed for dataset export.}
\usage{
getSeedForDatasetExport(seed, progress)
}
\arguments{
\item{seed}{original seed}

\item{progress}{simulation progress}
}
\value{
the seed value used to export the dataset
}
\description{
Get seed for dataset export.
}
