% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{getSeedForIteration}
\alias{getSeedForIteration}
\title{Get seed for iteration.}
\usage{
getSeedForIteration(seed, progress)
}
\arguments{
\item{seed}{original seed}

\item{progress}{simulation progress}
}
\value{
the seed value to be used for the given replicate number and iteration
}
\description{
Get seed for iteration.
}
