% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{getTimes}
\alias{getTimes}
\title{Get all distinct times across the events list.}
\usage{
getTimes(object)
}
\arguments{
\item{object}{any object}
}
\value{
numeric vector with all unique times, sorted
}
\description{
Get all distinct times across the events list.
}
