% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{getTimeVaryingCovariateNames}
\alias{getTimeVaryingCovariateNames}
\title{Get all time-varying covariate names.}
\usage{
getTimeVaryingCovariateNames(object)
}
\arguments{
\item{object}{any object}
}
\value{
character vector
}
\description{
Get all time-varying covariate names.
}
