% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardware_settings.R
\name{Hardware}
\alias{Hardware}
\title{Create hardware settings.}
\usage{
Hardware(cpu = 6, parallel = FALSE)
}
\arguments{
\item{cpu}{number of CPU's to use, default is 6}

\item{parallel}{enable parallel computing, logical value}
}
\value{
hardware settings
}
\description{
Create hardware settings.
}
