% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardware_settings.R
\docType{class}
\name{hardware_settings-class}
\alias{hardware_settings-class}
\title{Hardware settings class.}
\description{
Hardware settings class.
}
\section{Slots}{

\describe{
\item{\code{cpu}}{number of CPU's to use, default is 1}

\item{\code{slice_parallel}}{enable parallel computing for slices, default is FALSE}

\item{\code{slice_size}}{number of subjects per simulated slice, default is NULL (auto-configured by Campsis depending on the specified engine)}

\item{\code{dataset_parallel}}{enable parallelisation when exporting dataset into a table, default is FALSE}

\item{\code{dataset_slice_size}}{dataset slice size when exporting subjects to a table, default is 500. Only applicable if 'dataset_parallel' is enabled.}
}}

