% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_entry.R
\name{Infusion}
\alias{Infusion}
\title{Create an infusion bolus.}
\usage{
Infusion(time, amount, duration = NA, rate = NA, compartment = NA)
}
\arguments{
\item{time}{treatment time, numeric}

\item{amount}{amount to give as bolus, numeric}

\item{duration}{infusion duration}

\item{rate}{infusion rate}

\item{compartment}{compartment index, integer}
}
\value{
an observation
}
\description{
Create an infusion bolus.
}
