% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_entry.R
\name{Infusion}
\alias{Infusion}
\title{Create one or several infusion(s).}
\usage{
Infusion(
  time,
  amount,
  compartment = NULL,
  f = NULL,
  lag = NULL,
  duration = NULL,
  rate = NULL,
  ii = NULL,
  addl = NULL,
  wrap = FALSE,
  ref = NULL
)
}
\arguments{
\item{time}{treatment time(s), numeric value or vector. First treatment time if used together with ii and addl.}

\item{amount}{total amount to infuse, numeric}

\item{compartment}{compartment index or name to give the infusion(s). A vector of integers or names can be used for a complex model administration.}

\item{f}{fraction of infusion amount, distribution}

\item{lag}{infusion lag time, distribution}

\item{duration}{infusion duration, distribution}

\item{rate}{infusion rate, distribution}

\item{ii}{interdose interval, requires argument 'time' to be a single numeric value}

\item{addl}{number of additional doses, requires argument 'time' to be a single integer value}

\item{wrap}{if TRUE, the infusion wrapper will be stored as is in the dataset, otherwise,
it will be split into a list of infusions distinct in time. Default is FALSE.}

\item{ref}{wrapper reference, used to identify the wrapper in the dataset, single character value}
}
\value{
a single infusion or a list of infusions.
}
\description{
Create one or several infusion(s).
}
