% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_entry.R
\name{Infusion}
\alias{Infusion}
\title{Create an infusion.}
\usage{
Infusion(time, amount, compartment = NA)
}
\arguments{
\item{time}{treatment time, numeric}

\item{amount}{total amount to infuse, numeric}

\item{compartment}{compartment index, integer}
}
\value{
an infusion.
}
\description{
Create an infusion.
}
