% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_entry.R
\name{Infusion}
\alias{Infusion}
\title{Create an infusion.}
\usage{
Infusion(
  time,
  amount,
  compartment = NA,
  f = NULL,
  lag = NULL,
  duration = NULL,
  rate = NULL
)
}
\arguments{
\item{time}{treatment time, numeric}

\item{amount}{total amount to infuse, numeric}

\item{compartment}{compartment index, integer}

\item{f}{fraction of infusion amount, distribution}

\item{lag}{infusion lag time, distribution}

\item{duration}{infusion duration, distribution}

\item{rate}{infusion rate, distribution}
}
\value{
an infusion.
}
\description{
Create an infusion.
}
