% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_iov.R
\name{IOV}
\alias{IOV}
\title{Create IOV.}
\usage{
IOV(colname, distribution)
}
\arguments{
\item{colname}{name of the column that will be output in dataset}

\item{distribution}{distribution}
}
\value{
IOV
}
\description{
Create IOV.
}
