% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{is_empty_bootstrap}
\alias{is_empty_bootstrap}
\title{Is the given bootstrap empty.}
\usage{
is_empty_bootstrap(object)
}
\arguments{
\item{object}{bootstrap object}
}
\value{
logical value TRUE/FALSE
}
\description{
Is the given bootstrap empty.
}
\keyword{internal}
