% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_time.R
\name{LagTime}
\alias{LagTime}
\title{Create a lag time for the specified compartment.}
\usage{
LagTime(compartment, mean, variance)
}
\arguments{
\item{compartment}{compartment number}

\item{mean}{lag time mean value}

\item{variance}{lag time variance}
}
\value{
lag time
}
\description{
Create a lag time for the specified compartment.
}
