% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/time_vector.R,
%   R/treatment_entry.R, R/observations.R, R/dataset.R
\name{loadFromJSON}
\alias{loadFromJSON}
\alias{loadFromJSON,time_sequence,json_element-method}
\alias{loadFromJSON,bolus_wrapper,json_element-method}
\alias{loadFromJSON,infusion_wrapper,json_element-method}
\alias{loadFromJSON,observations,json_element-method}
\alias{loadFromJSON,dataset,json_element-method}
\alias{loadFromJSON,dataset,character-method}
\title{Fill-in S4 object from the JSON content.}
\usage{
loadFromJSON(object, json)

\S4method{loadFromJSON}{time_sequence,json_element}(object, json)

\S4method{loadFromJSON}{bolus_wrapper,json_element}(object, json)

\S4method{loadFromJSON}{infusion_wrapper,json_element}(object, json)

\S4method{loadFromJSON}{observations,json_element}(object, json)

\S4method{loadFromJSON}{dataset,json_element}(object, json)

\S4method{loadFromJSON}{dataset,character}(object, json)
}
\arguments{
\item{object}{pre-initiated S4 object}

\item{json}{JSON (usually a list)}
}
\value{
the loaded S4 object
}
\description{
Fill-in S4 object from the JSON content.
}
