% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{mergeTimeVaryingCovariates}
\alias{mergeTimeVaryingCovariates}
\title{Merge time-varying covariates into a single data frame. This last data frame
will be merged afterwards with all treatment and observation rows.}
\usage{
mergeTimeVaryingCovariates(covariates, ids_within_arm, arm_offset)
}
\arguments{
\item{covariates}{covariates, only time-varying covariates will be extracted}

\item{ids_within_arm}{ids within the current arm being sampled}

\item{arm_offset}{arm offset (in term of ID's)}
}
\value{
a data.frame
}
\description{
Merge time-varying covariates into a single data frame. This last data frame
will be merged afterwards with all treatment and observation rows.
}
\keyword{internal}
