% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{NormalDistribution}
\alias{NormalDistribution}
\title{Create a normal distribution.}
\usage{
NormalDistribution(mean, sd)
}
\arguments{
\item{mean}{mean}

\item{sd}{sd}
}
\value{
a normal distribution
}
\description{
Create a normal distribution.
}
