% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{obs_only}
\alias{obs_only}
\title{Filter Campsis output on observation rows.}
\usage{
obs_only(x)
}
\arguments{
\item{x}{data frame, Campsis output}
}
\value{
a data frame with the observation rows
}
\description{
Filter Campsis output on observation rows.
}
