% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{Observations}
\alias{Observations}
\title{Create a list of observations.}
\usage{
Observations(times, compartment = NA)
}
\arguments{
\item{compartment}{compartment index, integer}

\item{time}{observation times, numeric vector}
}
\value{
a list of observations, linked to the same compartment number
}
\description{
Create a list of observations.
}
