% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occasion.R
\name{Occasion}
\alias{Occasion}
\title{Create an occasion mapping. Occasions can only be defined at treatment times, via
the dose numbers.}
\usage{
Occasion(colname, values, doseNumbers)
}
\arguments{
\item{colname}{name of the column that will be output in dataset}

\item{values}{the occasion numbers, any integer vector}

\item{doseNumbers}{the related dose numbers, any integer vector of same length as 'values'}
}
\value{
occasion object
}
\description{
Create an occasion mapping. Occasions can only be defined at treatment times, via
the dose numbers.
}
