% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{on_ci}
\alias{on_ci}
\title{Check if the current session is on CI (e.g. GitHub actions).}
\usage{
on_ci()
}
\value{
logical value TRUE/FALSE
}
\description{
Check if the current session is on CI (e.g. GitHub actions).
}
\keyword{internal}
