% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun.R
\name{Outfun}
\alias{Outfun}
\title{Create a new output function}
\usage{
Outfun(
  fun = function(x, ...) {
     x
 },
  args = list(),
  packages = NULL,
  level = "replicate",
  fun_name = "default"
)
}
\arguments{
\item{fun}{function or purrr-style lambda formula, first argument 'x' must be the results}

\item{args}{extra arguments, named list}

\item{packages}{packages that must be loaded to execute the given function, character vector}

\item{level}{where to apply the output function, only 'replicate' is allowed since Campsis v1.9.0}

\item{fun_name}{name of the output function. Default is 'default'.}
}
\value{
an output function
}
\description{
Create a new output function
}
