% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun.R
\name{Outfun}
\alias{Outfun}
\title{Create a new output function}
\usage{
Outfun(
  fun = function(x, ...) {
     x
 },
  level = "scenario"
)
}
\arguments{
\item{fun}{function or purrr-style lambda formula}

\item{level}{either 'scenario' or 'replicate'. Default is 'scenario'.}
}
\value{
an output function
}
\description{
Create a new output function
}
