% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{ParameterDistribution}
\alias{ParameterDistribution}
\title{Create a parameter distribution.}
\usage{
ParameterDistribution(thetaName, etaName = character(0))
}
\arguments{
\item{name}{covariate name, character}

\item{value}{covariate value, numeric}
}
\value{
a covariate
}
\description{
Create a parameter distribution.
}
