% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{ParameterDistribution}
\alias{ParameterDistribution}
\title{Create a parameter distribution.}
\usage{
ParameterDistribution(thetaName, etaName = character(0))
}
\arguments{
\item{thetaName}{corresponding THETA name, character}

\item{etaName}{corresponding ETA name, character (character(0) if no ETA)}
}
\value{
a parameter distribution
}
\description{
Create a parameter distribution.
}
