% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{pi}
\alias{pi}
\title{Compute the prediction interval summary over time.}
\usage{
pi(x, output, scenarios = NULL, level = 0.9)
}
\arguments{
\item{x}{data frame}

\item{output}{variable to show}

\item{scenarios}{scenarios}

\item{level}{PI level, default is 0.9 (90% PI)}
}
\description{
Compute the prediction interval summary over time.
}
