% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_processing.R
\name{PI}
\alias{PI}
\title{Compute the prediction interval summary over time.}
\usage{
PI(x, output, scenarios = NULL, level = 0.9, gather = TRUE)
}
\arguments{
\item{x}{data frame}

\item{output}{variable to show, character value}

\item{scenarios}{scenarios, character vector, NULL is default}

\item{level}{PI level, default is 0.9 (90\% PI)}

\item{gather}{FALSE: med, low & up columns, TRUE: metric column}
}
\value{
summary
}
\description{
Compute the prediction interval summary over time.
}
