% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_processing.R
\name{PI}
\alias{PI}
\title{Compute the prediction interval summary over time.}
\usage{
PI(x, output, scenarios = NULL, level = 0.9)
}
\arguments{
\item{x}{data frame}

\item{output}{variable to show}

\item{scenarios}{scenarios}

\item{level}{PI level, default is 0.9 (90\% PI)}
}
\value{
summary
}
\description{
Compute the prediction interval summary over time.
}
