% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun.R
\name{PIOutfun}
\alias{PIOutfun}
\title{Create a prediction interval output function}
\usage{
PIOutfun(
  variable,
  strata = getDefaultStrata(),
  level = 0.9,
  fun_name = sprintf("PI_\%s_\%i\%\%", paste0(variable, collapse = "_"), round(level *
    100))
)
}
\arguments{
\item{variable}{variable(s) used to compute the prediction interval, character vector}

\item{strata}{named vector with the strata to use, default is c(SCENARIO="all", ARM="all")}

\item{level}{PI level, default is 0.9 (90\% PI)}

\item{fun_name}{name of the output function. Default is 'pi_<variable>_<level>pc'.}
}
\value{
a pi_outfun object
}
\description{
Create a prediction interval output function
}
