% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfun.R
\name{PIOutfun}
\alias{PIOutfun}
\title{Create a prediction interval output function}
\usage{
PIOutfun(variable, strata = getDefaultStrata(), level = 0.9, gather = TRUE)
}
\arguments{
\item{variable}{variable(s) used to compute the prediction interval, character vector}

\item{strata}{named vector with the strata to use, default is c(SCENARIO="all", ARM="all")}

\item{level}{PI level, default is 0.9 (90\% PI)}

\item{gather}{FALSE: med, low & up columns, TRUE: metric column}
}
\value{
a pi_outfun object
}
\description{
Create a prediction interval output function
}
