% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocess_events}
\alias{preprocess_events}
\title{Pre-process events.}
\usage{
preprocess_events(events)
}
\arguments{
\item{events}{interruption events}
}
\description{
Pre-process events.
}
\keyword{internal}
