% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocess_ids}
\alias{preprocess_ids}
\title{Preprocess subjects ID's.}
\usage{
preprocess_ids(dataset)
}
\arguments{
\item{dataset}{current dataset, data frame form}
}
\value{
list of consecutive ID's
}
\description{
Preprocess subjects ID's.
}
\keyword{internal}
