% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocess_outfun}
\alias{preprocess_outfun}
\title{Pre-process outfun argument.}
\usage{
preprocess_outfun(outfun)
}
\arguments{
\item{outfun}{function, lambda formula, 'outfun' or 'outfuns' object}
}
\value{
an 'outfuns' object in any case
}
\description{
Pre-process outfun argument.
}
\keyword{internal}
