% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocess_outvars}
\alias{preprocess_outvars}
\title{Preprocess 'outvars' argument. 'Outvars' is a character vector which tells
CAMPSIS the mandatory columns to keep in the output dataframe.}
\usage{
preprocess_outvars(outvars)
}
\arguments{
\item{outvars}{character vector or function}
}
\value{
outvars
}
\description{
Preprocess 'outvars' argument. 'Outvars' is a character vector which tells
CAMPSIS the mandatory columns to keep in the output dataframe.
}
\keyword{internal}
