% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocess_scenarios}
\alias{preprocess_scenarios}
\title{Pre-process scenarios.}
\usage{
preprocess_scenarios(scenarios)
}
\arguments{
\item{scenarios}{scenarios}
}
\description{
Pre-process scenarios.
}
\keyword{internal}
