% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocessEvents}
\alias{preprocessEvents}
\title{Pre-process events.}
\usage{
preprocessEvents(events)
}
\arguments{
\item{events}{interruption events}
}
\description{
Pre-process events.
}
\keyword{internal}
