% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocessNocb}
\alias{preprocessNocb}
\title{Preprocess 'nocb' argument.}
\usage{
preprocessNocb(nocb, dest)
}
\arguments{
\item{nocb}{nocb argument, logical value}

\item{dest}{destination engine}
}
\value{
user value, if not specified, return TRUE for mrgsolve and FALSE for RxODE
}
\description{
Preprocess 'nocb' argument.
}
\keyword{internal}
