% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocessOutfun}
\alias{preprocessOutfun}
\title{Pre-process outfun argument.}
\usage{
preprocessOutfun(outfun)
}
\arguments{
\item{fun}{function or lambda formula}
}
\value{
an output function
}
\description{
Pre-process outfun argument.
}
\keyword{internal}
