% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{preprocessOutfun}
\alias{preprocessOutfun}
\title{Preprocess 'outfun' argument. Outfun is a function (or predicate) that will be
call on the resulting output (at end of replicate simulation).}
\usage{
preprocessOutfun(outfun)
}
\arguments{
\item{outfun}{function or lambda formula, may be NULL}
}
\value{
outfun
}
\description{
Preprocess 'outfun' argument. Outfun is a function (or predicate) that will be
call on the resulting output (at end of replicate simulation).
}
\keyword{internal}
