% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{preprocessOutfun}
\alias{preprocessOutfun}
\title{Preprocess 'outvars' argument. Outvars is either the columns to keep from the 
returned dataframe or the function to apply on this dataframe.}
\usage{
preprocessOutfun(outfun)
}
\arguments{
\item{outvars}{character vector or function}
}
\value{
outvars
}
\description{
Preprocess 'outvars' argument. Outvars is either the columns to keep from the 
returned dataframe or the function to apply on this dataframe.
}
