% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocessOutvars}
\alias{preprocessOutvars}
\title{Preprocess 'outvars' argument. Outvars is a character vector which tells
pmxsim the mandatory columns to keep in the output dataframe.}
\usage{
preprocessOutvars(outvars)
}
\arguments{
\item{outvars}{character vector or function}
}
\value{
outvars
}
\description{
Preprocess 'outvars' argument. Outvars is a character vector which tells
pmxsim the mandatory columns to keep in the output dataframe.
}
\keyword{internal}
