% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{preprocessReplicates}
\alias{preprocessReplicates}
\title{Preprocess 'replicates' argument.}
\usage{
preprocessReplicates(replicates, model)
}
\arguments{
\item{replicates}{number of replicates}

\item{model}{Campsis model (class 'campsis_model' or 'replicated_campsis_model')}
}
\value{
number of replicates to simulate
}
\description{
Preprocess 'replicates' argument.
}
\keyword{internal}
