% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{preserve_column_levels}
\alias{preserve_column_levels}
\title{Preserve Existing Column Value Order as Factor Levels}
\usage{
preserve_column_levels(x, cols)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{cols}{A character vector of column names to convert.}
}
\value{
A data frame with updated factor columns.
}
\description{
Converts target columns into factors using their current unique row appearance
order as the factor levels.
}
