% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_preprocess.R
\name{process_drop_others}
\alias{process_drop_others}
\title{Process 'DROP_OTHERS'.}
\usage{
process_drop_others(x, outvars = character(0), drop_others)
}
\arguments{
\item{x}{the current data frame}

\item{outvars}{variables to keep}

\item{drop_others}{logical value}
}
\value{
processed data frame
}
\description{
Process 'DROP_OTHERS'.
}
\keyword{internal}
