% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{processAsCovariate}
\alias{processAsCovariate}
\title{Process distribution-derived objects as covariates.}
\usage{
processAsCovariate(list, iiv, rxmod, ids)
}
\arguments{
\item{list}{list of distribution-derived objects}

\item{iiv}{IIV pre-computed data frame}

\item{rxmod}{prepared data for RxODE}

\item{ids}{subject ID's being simulated in arm}
}
\value{
a list of covariates that can be easily processed
}
\description{
Process distribution-derived objects as covariates.
}
