% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{processReturnedDataframe}
\alias{processReturnedDataframe}
\title{Process the data frame returned by mrgsolve or RxODE.}
\usage{
processReturnedDataframe(x, outvars = NULL, outfun = NULL)
}
\arguments{
\item{x}{the current data frame}

\item{outvars}{variables to keep}

\item{outfun}{processing function}
}
\value{
processed data frame
}
\description{
Process the data frame returned by mrgsolve or RxODE.
}
