% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{processSimulateArguments}
\alias{processSimulateArguments}
\title{Preprocess arguments of the simulate method.}
\usage{
processSimulateArguments(model, dataset, dest, outvars, dosing, settings, ...)
}
\arguments{
\item{model}{CAMPSIS model}

\item{dataset}{dataset, data.frame form}

\item{dest}{destination engine}

\item{outvars}{outvars}

\item{dosing}{add dosing information, logical value}

\item{settings}{simulation settings}

\item{...}{all other arguments}
}
\value{
a simulation configuration
}
\description{
Preprocess arguments of the simulate method.
}
\keyword{internal}
