% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{reorder_columns}
\alias{reorder_columns}
\title{Reorder output columns.}
\usage{
reorder_columns(results, dosing)
}
\arguments{
\item{results}{RxODE/mrgsolve output}

\item{dosing}{dosing information, logical value}
}
\value{
reordered dataframe
}
\description{
Reorder output columns.
}
\keyword{internal}
