% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/distribution.R,
%   R/treatment_entry.R, R/observations.R, R/covariate.R, R/bootstrap.R,
%   R/parameter_uncertainty.R
\name{sample}
\alias{sample}
\alias{sample,constant_distribution,integer-method}
\alias{sample,fixed_distribution,integer-method}
\alias{sample,function_distribution,integer-method}
\alias{sample,bootstrap_distribution,integer-method}
\alias{sample,bolus,integer-method}
\alias{sample,infusion,integer-method}
\alias{sample,observations,integer-method}
\alias{sample,covariate,integer-method}
\alias{sample,bootstrap,integer-method}
\alias{sample,campsis_model,integer-method}
\title{Sample generic object.}
\usage{
sample(object, n, ...)

\S4method{sample}{constant_distribution,integer}(object, n)

\S4method{sample}{fixed_distribution,integer}(object, n)

\S4method{sample}{function_distribution,integer}(object, n)

\S4method{sample}{bootstrap_distribution,integer}(object, n)

\S4method{sample}{bolus,integer}(object, n, ...)

\S4method{sample}{infusion,integer}(object, n, ...)

\S4method{sample}{observations,integer}(object, n, ...)

\S4method{sample}{covariate,integer}(object, n)

\S4method{sample}{bootstrap,integer}(object, n)

\S4method{sample}{campsis_model,integer}(object, n)
}
\arguments{
\item{object}{generic object}

\item{n}{number of samples required}

\item{...}{extra arguments}
}
\value{
sampling result
}
\description{
Sample generic object.
}
