% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{sample_covariates_list}
\alias{sample_covariates_list}
\title{Sample covariates list.}
\usage{
sample_covariates_list(covariates, ids_within_arm, subset)
}
\arguments{
\item{covariates}{list of covariates to sample}

\item{ids_within_arm}{ids within the current arm being sampled}

\item{subset}{take subset of original values because export is parallelised}
}
\value{
a dataframe of n rows, 1 column per covariate
}
\description{
Sample covariates list.
}
\keyword{internal}
