% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{sampleCovariatesList}
\alias{sampleCovariatesList}
\title{Sample covariates list.}
\usage{
sampleCovariatesList(covariates, n)
}
\arguments{
\item{covariates}{list of covariates to sample}

\item{n}{number of desired samples}
}
\value{
a dataframe of n rows, 1 column per covariate
}
\description{
Sample covariates list.
}
