% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{sampleTimeVaryingCovariates}
\alias{sampleTimeVaryingCovariates}
\title{Sample time-varying covariates.}
\usage{
sampleTimeVaryingCovariates(object, armID, needsDV)
}
\arguments{
\item{covariates}{time-varying covariates}

\item{ids}{simulation ID's}
}
\value{
a data.frame
}
\description{
Sample time-varying covariates.
}
\keyword{internal}
