% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter plot (or X vs Y plot).}
\usage{
scatterPlot(x, variable, colour = NULL, time = NULL)
}
\arguments{
\item{x}{data frame}

\item{variable}{the 2 variables to show, character vector}

\item{colour}{variable(s) to colour}

\item{time}{the time to look at those 2 variables, if NULL, min time is used (usually 0)}
}
\value{
a ggplot object
}
\description{
Scatter plot (or X vs Y plot).
}
\keyword{internal}
