% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\docType{class}
\name{scenario-class}
\alias{scenario-class}
\title{Scenario class.}
\description{
Scenario class.
}
\section{Slots}{

\describe{
\item{\code{name}}{scenario name, single character string}

\item{\code{model}}{either a CAMPSIS model, a function or lambda-style formula}

\item{\code{dataset}}{either a CAMPSIS dataset, a function or lambda-style formula}
}}

