% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{Scenario}
\alias{Scenario}
\title{Create an scenario.}
\usage{
Scenario(name = NULL, model = NULL, dataset = NULL)
}
\arguments{
\item{name}{scenario name, single character string}

\item{model}{either a CAMPSIS model, a function or lambda-style formula}

\item{dataset}{either a CAMPSIS dataset, a function or lambda-style formula}
}
\value{
a new scenario
}
\description{
Create an scenario.
}
