% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{Scenarios}
\alias{Scenarios}
\title{Create a list of scenarios.}
\usage{
Scenarios(json = NULL)
}
\arguments{
\item{json}{path to JSON scenarios file or JSON content in string form}
}
\value{
a scenarios object
}
\description{
Create a list of scenarios.
}
