% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/arm.R, R/dataset.R
\name{set_label}
\alias{set_label}
\alias{set_label,arm,character-method}
\alias{set_label,dataset,character-method}
\title{Set the label.}
\usage{
set_label(object, x)

\S4method{set_label}{arm,character}(object, x)

\S4method{set_label}{dataset,character}(object, x)
}
\arguments{
\item{object}{any object that has a label}

\item{x}{the new label}
}
\value{
the updated object
}
\description{
Set the label.
}
