% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{set_seed}
\alias{set_seed}
\title{Set the seed. The goal of this method is to centralize all calls to
the R method 'set.seed'.}
\usage{
set_seed(seed)
}
\arguments{
\item{seed}{seed value, not NULL}
}
\description{
Set the seed. The goal of this method is to centralize all calls to
the R method 'set.seed'.
}
\keyword{internal}
