% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{setLabel}
\alias{setLabel}
\title{Set the label.}
\usage{
setLabel(object, x)
}
\arguments{
\item{object}{any object that has a label}

\item{x}{the new label}
}
\value{
the updated object
}
\description{
Set the label.
}
