% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{setSubjects}
\alias{setSubjects}
\title{Set the number of subjects.}
\usage{
setSubjects(object, x)
}
\arguments{
\item{object}{any object}

\item{x}{the new number of subjects}
}
\value{
the updated object
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`setSubjects()` is deprecated in favor of `set_subjects()`.
}
