% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/arm.R, R/dataset.R
\name{setSubjects}
\alias{setSubjects}
\alias{setSubjects,arm,integer-method}
\alias{setSubjects,dataset,integer-method}
\title{Set the number of subjects.}
\usage{
setSubjects(object, x)

\S4method{setSubjects}{arm,integer}(object, x)

\S4method{setSubjects}{dataset,integer}(object, x)
}
\arguments{
\item{object}{any object}

\item{x}{the new number of subjects}
}
\value{
the updated object
}
\description{
Set the number of subjects.
}
