% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{setSubjects}
\alias{setSubjects}
\title{Set the number of subjects.}
\usage{
setSubjects(object, x)
}
\arguments{
\item{object}{any object}

\item{x}{the new number of subjects}
}
\value{
the updated object
}
\description{
Set the number of subjects.
}
