% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{Settings}
\alias{Settings}
\title{Create advanced simulation settings.}
\usage{
Settings(..., json = NULL)
}
\arguments{
\item{...}{any user-required settings: see ?Hardware, ?Solver, ?NOCB, ?Declare, ?Progress or ?AutoReplicationSettings}

\item{json}{path to JSON settings file or JSON content in string form}
}
\value{
advanced simulation settings
}
\description{
Create advanced simulation settings.
}
