% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{Settings}
\alias{Settings}
\title{Create advanced simulation settings.}
\usage{
Settings(..., cpu = 1, parallel = FALSE)
}
\arguments{
\item{...}{other settings like hardware settings, solver settings, NOCB settings, etc.}

\item{cpu}{number of CPU's to use, default is 1}

\item{parallel}{enable parallel computing, logical value}
}
\value{
advanced simulation settings
}
\description{
Create advanced simulation settings.
}
