% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{Settings}
\alias{Settings}
\title{Create advanced simulation settings.}
\usage{
Settings(...)
}
\arguments{
\item{...}{any user-required settings: see ?Hardware, ?Solver, ?NOCB or ?Declare settings}
}
\value{
advanced simulation settings
}
\description{
Create advanced simulation settings.
}
