% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_setup.R
\name{setup_plan_sequential}
\alias{setup_plan_sequential}
\title{Setup plan as sequential (i.e. no parallelisation).}
\usage{
setup_plan_sequential()
}
\value{
nothing
}
\description{
Setup plan as sequential (i.e. no parallelisation).
}
