% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_setup.R
\name{setupPlanDefault}
\alias{setupPlanDefault}
\title{Setup plan for the given simulation or hardware settings.}
\usage{
setupPlanDefault(object)
}
\arguments{
\item{object}{simulation or hardware settings}
}
\value{
nothing
}
\description{
Setup plan for the given simulation or hardware settings.
}
