% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_setup.R
\name{setupPlanDefault}
\alias{setupPlanDefault}
\title{Setup plan for the given hardware settings.}
\usage{
setupPlanDefault(hardware)
}
\arguments{
\item{hardware}{hardware settings object}
}
\value{
nothing
}
\description{
Setup plan for the given hardware settings.
}
