% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{setupPlanSequential}
\alias{setupPlanSequential}
\title{Setup plan as sequential (i.e. no parallelisation).}
\usage{
setupPlanSequential()
}
\value{
nothing
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`setupPlanSequential()` is deprecated in favor of `setup_plan_sequential()`.
}
