% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_setup.R
\name{setupPlanSequential}
\alias{setupPlanSequential}
\title{Setup plan as sequential (i.e. no parallelisation).}
\usage{
setupPlanSequential()
}
\value{
nothing
}
\description{
Setup plan as sequential (i.e. no parallelisation).
}
