% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaded_plot.R
\name{shaded_plot}
\alias{shaded_plot}
\title{Shaded plot / prediction interval plot (S3 generic).}
\usage{
shaded_plot(x, ...)
}
\arguments{
\item{x}{a Campsis output object}

\item{...}{additional arguments passed to the method}
}
\value{
a ggplot object
}
\description{
Shaded plot / prediction interval plot (S3 generic).
}
