% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\alias{simulate,replicated_campsis_model,dataset,character,events,scenarios,function,character,output_functions,integer,integer,logical,simulation_settings-method}
\alias{simulate,campsis_model,dataset,character,events,scenarios,function,character,output_functions,integer,integer,logical,simulation_settings-method}
\alias{simulate,campsis_model,tbl_df,character,events,scenarios,function,character,output_functions,integer,integer,logical,simulation_settings-method}
\alias{simulate,campsis_model,data.frame,character,events,scenarios,function,character,output_functions,integer,integer,logical,simulation_settings-method}
\alias{simulate,campsis_model,tbl_df,rxode_engine,events,scenarios,function,character,output_functions,integer,integer,logical,simulation_settings-method}
\alias{simulate,campsis_model,tbl_df,mrgsolve_engine,events,scenarios,function,character,output_functions,integer,integer,logical,simulation_settings-method}
\title{Simulate function.}
\usage{
simulate(
  model,
  dataset,
  dest = NULL,
  events = NULL,
  scenarios = NULL,
  tablefun = NULL,
  outvars = NULL,
  outfun = NULL,
  seed = NULL,
  replicates = 1,
  dosing = FALSE,
  settings = NULL
)

\S4method{simulate}{replicated_campsis_model,dataset,character,events,scenarios,function,character,output_functions,integer,integer,logical,simulation_settings}(
  model,
  dataset,
  dest = NULL,
  events = NULL,
  scenarios = NULL,
  tablefun = NULL,
  outvars = NULL,
  outfun = NULL,
  seed = NULL,
  replicates = 1,
  dosing = FALSE,
  settings = NULL
)

\S4method{simulate}{campsis_model,dataset,character,events,scenarios,function,character,output_functions,integer,integer,logical,simulation_settings}(
  model,
  dataset,
  dest = NULL,
  events = NULL,
  scenarios = NULL,
  tablefun = NULL,
  outvars = NULL,
  outfun = NULL,
  seed = NULL,
  replicates = 1,
  dosing = FALSE,
  settings = NULL
)

\S4method{simulate}{campsis_model,tbl_df,character,events,scenarios,function,character,output_functions,integer,integer,logical,simulation_settings}(
  model,
  dataset,
  dest = NULL,
  events = NULL,
  scenarios = NULL,
  tablefun = NULL,
  outvars = NULL,
  outfun = NULL,
  seed = NULL,
  replicates = 1,
  dosing = FALSE,
  settings = NULL
)

\S4method{simulate}{campsis_model,data.frame,character,events,scenarios,function,character,output_functions,integer,integer,logical,simulation_settings}(
  model,
  dataset,
  dest = NULL,
  events = NULL,
  scenarios = NULL,
  tablefun = NULL,
  outvars = NULL,
  outfun = NULL,
  seed = NULL,
  replicates = 1,
  dosing = FALSE,
  settings = NULL
)

\S4method{simulate}{campsis_model,tbl_df,rxode_engine,events,scenarios,function,character,output_functions,integer,integer,logical,simulation_settings}(
  model,
  dataset,
  dest = NULL,
  events = NULL,
  scenarios = NULL,
  tablefun = NULL,
  outvars = NULL,
  outfun = NULL,
  seed = NULL,
  replicates = 1,
  dosing = FALSE,
  settings = NULL
)

\S4method{simulate}{campsis_model,tbl_df,mrgsolve_engine,events,scenarios,function,character,output_functions,integer,integer,logical,simulation_settings}(
  model,
  dataset,
  dest = NULL,
  events = NULL,
  scenarios = NULL,
  tablefun = NULL,
  outvars = NULL,
  outfun = NULL,
  seed = NULL,
  replicates = 1,
  dosing = FALSE,
  settings = NULL
)
}
\arguments{
\item{model}{generic CAMPSIS model}

\item{dataset}{CAMPSIS dataset or 2-dimensional table}

\item{dest}{destination simulation engine, default is 'RxODE'}

\item{events}{interruption events}

\item{scenarios}{list of scenarios to be simulated}

\item{tablefun}{function or lambda formula to apply on exported 2-dimensional dataset}

\item{outvars}{variables to output in resulting dataframe}

\item{outfun}{an output function to apply on the simulation results. Type ?Outfun for more info.}

\item{seed}{seed value}

\item{replicates}{number of replicates, default is 1}

\item{dosing}{output dosing information, default is FALSE}

\item{settings}{advanced simulation settings}
}
\value{
dataframe with all results
}
\description{
Simulate function.
}
