% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{simulate}
\alias{simulate}
\title{#' Convert generic object according to given configuration.
#' 
#' @param object generic object
#' @param config specified configuration
#' @return conversion output
#' @export
convert <- function(object, config) {
  stop("No default function is provided")
}}
\usage{
simulate(model, dataset, dest, ...)
}
\arguments{
\item{model}{generic model}

\item{dataset}{generic dataset}

\item{dest}{destination simulation engine}

\item{...}{optional arguments}
}
\value{
specific object depending on given destination
}
\description{
setGeneric("convert", function(object, config) {
  standardGeneric("convert")
})
Simulate function.
}
