% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{simulate}
\alias{simulate}
\title{Simulate function.}
\usage{
simulate(model, dataset, dest, ...)
}
\arguments{
\item{model}{generic model}

\item{dataset}{generic dataset}

\item{dest}{destination simulation engine}

\item{...}{optional arguments}
}
\value{
specific object depending on given destination
}
\description{
Simulate function.
}
