% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\alias{simulate,pmx_model,dataset,character,events,function,character,function,integer,integer,logical-method}
\alias{simulate,pmx_model,data.frame,character,events,function,character,function,integer,integer,logical-method}
\alias{simulate,pmx_model,data.frame,rxode_engine,events,function,character,function,integer,integer,logical-method}
\alias{simulate,pmx_model,data.frame,mrgsolve_engine,events,function,character,function,integer,integer,logical-method}
\title{Simulate function.}
\usage{
simulate(
  model,
  dataset,
  dest = NULL,
  events = NULL,
  tablefun = NULL,
  outvars = NULL,
  outfun = NULL,
  seed = NULL,
  replicates = 1,
  nocb = NULL,
  ...
)

\S4method{simulate}{pmx_model,dataset,character,events,`function`,character,`function`,integer,integer,logical}(
  model,
  dataset,
  dest = NULL,
  events = NULL,
  tablefun = NULL,
  outvars = NULL,
  outfun = NULL,
  seed = NULL,
  replicates = 1,
  nocb = NULL,
  ...
)

\S4method{simulate}{pmx_model,data.frame,character,events,`function`,character,`function`,integer,integer,logical}(
  model,
  dataset,
  dest = NULL,
  events = NULL,
  tablefun = NULL,
  outvars = NULL,
  outfun = NULL,
  seed = NULL,
  replicates = 1,
  nocb = NULL,
  ...
)

\S4method{simulate}{pmx_model,data.frame,rxode_engine,events,`function`,character,`function`,integer,integer,logical}(
  model,
  dataset,
  dest = NULL,
  events = NULL,
  tablefun = NULL,
  outvars = NULL,
  outfun = NULL,
  seed = NULL,
  replicates = 1,
  nocb = NULL,
  ...
)

\S4method{simulate}{pmx_model,data.frame,mrgsolve_engine,events,`function`,character,`function`,integer,integer,logical}(
  model,
  dataset,
  dest = NULL,
  events = NULL,
  tablefun = NULL,
  outvars = NULL,
  outfun = NULL,
  seed = NULL,
  replicates = 1,
  nocb = NULL,
  ...
)
}
\arguments{
\item{model}{generic PMX model}

\item{dataset}{PMX dataset or 2-dimensional table}

\item{dest}{destination simulation engine, default is 'RxODE'}

\item{events}{interruption events}

\item{tablefun}{function or lambda formula to apply on exported 2-dimensional dataset}

\item{outvars}{variables to output in resulting dataframe}

\item{outfun}{function or lambda formula to apply on resulting dataframe after each replicate}

\item{seed}{seed value}

\item{replicates}{number of replicates, default is 1}

\item{nocb}{next-observation carried backward mode (NOCB), default value is TRUE for mrgsolve, FALSE for RxODE}

\item{...}{optional arguments like declare}
}
\value{
dataframe with all results
}
\description{
Simulate function.
}
