% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateDelegate}
\alias{simulateDelegate}
\title{Simulation delegate.}
\usage{
simulateDelegate(
  model,
  dataset,
  dest,
  tablefun,
  outvars,
  outfun,
  seed,
  replicates,
  ...
)
}
\arguments{
\item{model}{generic PMX model}

\item{dataset}{PMX dataset or 2-dimensional table}

\item{dest}{destination simulation engine, default is 'RxODE'}

\item{tablefun}{function or lambda formula to apply on exported 2-dimensional dataset}

\item{outvars}{variables to output in resulting dataframe}

\item{outfun}{function or lambda formula to apply on resulting dataframe after each replicate}

\item{seed}{seed value}

\item{replicates}{number of replicates, default is 1}

\item{...}{optional arguments like declare}
}
\description{
Simulation delegate.
}
\keyword{internal}
