% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_progress.R
\docType{class}
\name{simulation_progress-class}
\alias{simulation_progress-class}
\title{Simulation progress class.}
\arguments{
\item{replicates}{total number of replicates to simulate}

\item{scenarios}{total number of scenarios to simulate}

\item{iterations}{total number of iterations to simulate}

\item{slices}{total number of slices to simulate}

\item{replicate}{current replicate number being simulated}

\item{scenario}{current scenario number being simulated}

\item{iteration}{current iteration number being simulated}

\item{slice}{current slice number being simulated}
}
\description{
Simulation progress class.
}
