% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\docType{class}
\name{simulation_settings-class}
\alias{simulation_settings-class}
\title{Simulation settings class.}
\description{
Simulation settings class.
}
\section{Slots}{

\describe{
\item{\code{hardware}}{hardware settings object}

\item{\code{solver}}{solver settings object}

\item{\code{nocb}}{NOCB settings object}
}}

