% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_progress.R
\name{SimulationProgress}
\alias{SimulationProgress}
\title{Create a simulation progress object.}
\usage{
SimulationProgress(replicates = 1, scenarios = 1)
}
\arguments{
\item{replicates}{total number of replicates to simulate}

\item{scenarios}{total number of scenarios to simulate}
}
\value{
a progress bar
}
\description{
Create a simulation progress object.
}
