% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_settings.R
\name{Solver}
\alias{Solver}
\title{Create solver settings.}
\usage{
Solver(
  atol = 1e-08,
  rtol = 1e-08,
  hmax = NA,
  maxsteps = 70000L,
  method = "liblsoda"
)
}
\value{
solver settings
}
\description{
Create solver settings.
}
\section{Slots}{

\describe{
\item{\code{atol}}{absolute solver tolerance}

\item{\code{rtol}}{relative solver tolerance}

\item{\code{hmax}}{limit how big a solver step can be}

\item{\code{maxsteps}}{max steps between 2 integration times (e.g. when observations records are far apart)}

\item{\code{method}}{solver method}
}}

