% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_settings.R
\docType{class}
\name{solver_settings-class}
\alias{solver_settings-class}
\title{Solver settings class.
See ?mrgsolve::update.
See ?rxode2::rxSolve.}
\description{
Solver settings class.
See ?mrgsolve::update.
See ?rxode2::rxSolve.
}
\section{Slots}{

\describe{
\item{\code{atol}}{absolute solver tolerance, default is 1e-08}

\item{\code{rtol}}{relative solver tolerance, default is 1e-08}

\item{\code{hmax}}{limit how big a solver step can be, default is NA}

\item{\code{maxsteps}}{max steps between 2 integration times (e.g. when observations records are far apart), default is 70000}

\item{\code{method}}{solver method, for RxODE/rxode2 only: 'liblsoda' (default), 'lsoda', 'dop853', 'indLin'. Mrgsolve's method is always 'lsoda'.}
}}

