% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_settings.R
\docType{class}
\name{solver_settings-class}
\alias{solver_settings-class}
\title{Solver settings class.
See ?mrgsolve::update.
See ?rxode2::rxSolve.}
\description{
Solver settings class.
See ?mrgsolve::update.
See ?rxode2::rxSolve.
}
\section{Slots}{

\describe{
\item{\code{atol}}{absolute solver tolerance}

\item{\code{rtol}}{relative solver tolerance}

\item{\code{hmax}}{limit how big a solver step can be}

\item{\code{maxsteps}}{max steps between 2 integration times (e.g. when observations records are far apart)}

\item{\code{method}}{solver method}
}}

