% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghetti_plot.R
\name{spaghetti_plot}
\alias{spaghetti_plot}
\title{Spaghetti plot (S3 generic).}
\usage{
spaghetti_plot(x, ...)
}
\arguments{
\item{x}{a Campsis output object}

\item{...}{additional arguments passed to the method}
}
\value{
a ggplot object
}
\description{
Spaghetti plot (S3 generic).
}
