% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{spaghettiPlot}
\alias{spaghettiPlot}
\title{Spaghetti plot.}
\usage{
spaghettiPlot(x, variable, colour = NULL)
}
\arguments{
\item{x}{data frame}

\item{variable}{variable to show}

\item{colour}{variable(s) to colour}
}
\value{
plot
}
\description{
Spaghetti plot.
}
