% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utilities.R
\name{standardise_time}
\alias{standardise_time}
\title{Standardise time to hours.}
\usage{
standardise_time(x, unit)
}
\arguments{
\item{x}{numeric time vector}

\item{unit}{unit of x, single character value}
}
\value{
numeric vector with the times converted to hours
}
\description{
Standardise time to hours.
}
