% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{TimeVaryingCovariate}
\alias{TimeVaryingCovariate}
\title{Create a time-varying covariate. These covariates can be modified further in the
interruption events.}
\usage{
TimeVaryingCovariate(name, distribution)
}
\arguments{
\item{name}{covariate name, character}

\item{distribution}{covariate distribution at time 0}
}
\value{
a time-varying covariate
}
\description{
Create a time-varying covariate. These covariates can be modified further in the
interruption events.
}
