% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{TimeVaryingCovariate}
\alias{TimeVaryingCovariate}
\title{Create a time-varying covariate.}
\usage{
TimeVaryingCovariate(name, table)
}
\arguments{
\item{name}{covariate name, character}

\item{table}{data.frame, must contain the mandatory columns 'TIME' and 'VALUE'.
An 'ID' column may also be specified. In that case, ID's between 1 and the
max number of subjects in the dataset/arm can be used.}
}
\value{
a time-varying covariate
}
\description{
Create a time-varying covariate.
}
