% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{to_explicit_distribution}
\alias{to_explicit_distribution}
\title{Convert user-given distribution to an explicit CAMPSIS distribution.
Passed distribution can be:
- a NULL value. In that case, it will be converted into an 'UndefinedDistribution'.
- a single numeric value. In that case, it will be converted into a 'ConstantDistribution'.
- a numeric vector. In that case, it will be converted into a 'FixedDistribution'.
- all available types of distribution. In this case, no conversion is applied.}
\usage{
to_explicit_distribution(distribution)
}
\arguments{
\item{distribution}{user-given distribution}
}
\value{
a distribution object
}
\description{
Convert user-given distribution to an explicit CAMPSIS distribution.
Passed distribution can be:
- a NULL value. In that case, it will be converted into an 'UndefinedDistribution'.
- a single numeric value. In that case, it will be converted into a 'ConstantDistribution'.
- a numeric vector. In that case, it will be converted into a 'FixedDistribution'.
- all available types of distribution. In this case, no conversion is applied.
}
\keyword{internal}
