% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment_entry.R, R/treatment.R,
%   R/arm.R, R/arms.R, R/dataset.R
\name{unwrap_treatment}
\alias{unwrap_treatment}
\alias{unwrap_treatment,bolus-method}
\alias{unwrap_treatment,infusion-method}
\alias{unwrap_treatment,bolus_wrapper-method}
\alias{unwrap_treatment,infusion_wrapper-method}
\alias{unwrap_treatment,treatment-method}
\alias{unwrap_treatment,arm-method}
\alias{unwrap_treatment,arms-method}
\alias{unwrap_treatment,dataset-method}
\title{Unwrap treatment.}
\usage{
unwrap_treatment(object)

\S4method{unwrap_treatment}{bolus}(object)

\S4method{unwrap_treatment}{infusion}(object)

\S4method{unwrap_treatment}{bolus_wrapper}(object)

\S4method{unwrap_treatment}{infusion_wrapper}(object)

\S4method{unwrap_treatment}{treatment}(object)

\S4method{unwrap_treatment}{arm}(object)

\S4method{unwrap_treatment}{arms}(object)

\S4method{unwrap_treatment}{dataset}(object)
}
\arguments{
\item{object}{any object}
}
\value{
updated object
}
\description{
Unwrap treatment.
}
