% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/treatment_entry.R, R/treatment.R,
%   R/arm.R, R/arms.R, R/dataset.R
\name{update_ii}
\alias{update_ii}
\alias{update_ii,bolus_wrapper,numeric,character-method}
\alias{update_ii,infusion_wrapper,numeric,character-method}
\alias{update_ii,bolus,numeric,character-method}
\alias{update_ii,infusion,numeric,character-method}
\alias{update_ii,treatment,numeric,character-method}
\alias{update_ii,arm,numeric,character-method}
\alias{update_ii,arms,numeric,character-method}
\alias{update_ii,dataset,numeric,character-method}
\title{Update the inter-dose interval (II).}
\usage{
update_ii(object, ii, ref = NULL)

\S4method{update_ii}{bolus_wrapper,numeric,character}(object, ii, ref = NULL)

\S4method{update_ii}{infusion_wrapper,numeric,character}(object, ii, ref = NULL)

\S4method{update_ii}{bolus,numeric,character}(object, ii, ref = NULL)

\S4method{update_ii}{infusion,numeric,character}(object, ii, ref = NULL)

\S4method{update_ii}{treatment,numeric,character}(object, ii, ref = NULL)

\S4method{update_ii}{arm,numeric,character}(object, ii, ref = NULL)

\S4method{update_ii}{arms,numeric,character}(object, ii, ref = NULL)

\S4method{update_ii}{dataset,numeric,character}(object, ii, ref = NULL)
}
\arguments{
\item{object}{generic object}

\item{ii}{new inter-dose interval}

\item{ref}{reference treatment name}
}
\value{
updated object
}
\description{
Update the inter-dose interval (II).
}
