% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_plot.R
\name{vpc_plot}
\alias{vpc_plot}
\title{VPC plot (S3 generic).}
\usage{
vpc_plot(x, ...)
}
\arguments{
\item{x}{a Campsis output object}

\item{...}{additional arguments passed to the method}
}
\value{
a ggplot object
}
\description{
VPC plot (S3 generic).
}
