% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{vpcPlot}
\alias{vpcPlot}
\title{VPC plot.}
\usage{
vpcPlot(x, strata = NULL, level = 0.9, alpha = 0.15)
}
\arguments{
\item{x}{data frame, output of CAMPSIS with replicates}

\item{strata}{named vector with the strata to use, default is c(SCENARIO="all", ARM="all").
Only columns that are actually present in \code{x} are used.}

\item{level}{PI level, default is 0.9 (90\% PI)}

\item{alpha}{alpha parameter (transparency) given to geom_ribbon}
}
\value{
a ggplot object
}
\description{
VPC plot.
}
