% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_plot.R
\name{vpcPlot}
\alias{vpcPlot}
\title{VPC plot (1 plot per scenario).}
\usage{
vpcPlot(x, scenarios = NULL, level = 0.9)
}
\arguments{
\item{x}{data frame, output of pmxsim with replicates}

\item{scenarios}{scenarios, character vector, NULL is default}

\item{level}{PI level, default is 0.9 (90\% PI)}
}
\value{
plot
}
\description{
VPC plot (1 plot per scenario).
}
